@@ -39,6 +39,7 @@ import android.text.style.TypefaceSpan;
 import android.util.SparseArray;
 import android.view.LayoutInflater;
 import android.view.Menu;
+import android.view.MenuInflater;
 import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
@@ -52,7 +53,7 @@ import androidx.annotation.Nullable;
 import androidx.appcompat.widget.SearchView;
 import androidx.constraintlayout.widget.ConstraintLayout;
 import androidx.coordinatorlayout.widget.CoordinatorLayout;
-import androidx.core.content.ContextCompat;
+import androidx.core.view.MenuProvider;
 import androidx.fragment.app.Fragment;
 import androidx.navigation.NavOptions;
 import androidx.recyclerview.widget.LinearLayoutManager;
@@ -93,7 +94,7 @@ import rikka.core.util.ResourceUtils;
 import rikka.material.app.LocaleDelegate;
 import rikka.recyclerview.RecyclerViewKt;
 
-public class ModulesFragment extends BaseFragment implements ModuleUtil.ModuleListener, RepoLoader.RepoListener {
+public class ModulesFragment extends BaseFragment implements ModuleUtil.ModuleListener, RepoLoader.RepoListener, MenuProvider {
     private static final PackageManager pm = App.getInstance().getPackageManager();
     private static final ModuleUtil moduleUtil = ModuleUtil.getInstance();
     private static final RepoLoader repoLoader = RepoLoader.getInstance();
@@ -191,20 +192,32 @@ public class ModulesFragment extends BaseFragment implements ModuleUtil.ModuleLi
     }
 
     @Override
-    public void onPrepareOptionsMenu(Menu menu) {
+    public void onPrepareMenu(Menu menu) {
         searchView = (SearchView) menu.findItem(R.id.menu_search).getActionView();
-        searchView.setOnQueryTextListener(searchListener);
-        searchView.addOnAttachStateChangeListener(new View.OnAttachStateChangeListener() {
-            @Override
-            public void onViewAttachedToWindow(View arg0) {
-                binding.appBar.setExpanded(false, true);
-            }
+        if (searchView != null) {
+            searchView.setOnQueryTextListener(searchListener);
+            searchView.addOnAttachStateChangeListener(new View.OnAttachStateChangeListener() {
+                @Override
+                public void onViewAttachedToWindow(@NonNull View arg0) {
+                    binding.appBar.setExpanded(false, true);
+                }
 
-            @Override
-            public void onViewDetachedFromWindow(View v) {
-            }
-        });
-        searchView.findViewById(androidx.appcompat.R.id.search_edit_frame).setLayoutDirection(View.LAYOUT_DIRECTION_INHERIT);
+                @Override
+                public void onViewDetachedFromWindow(@NonNull View v) {
+                }
+            });
+            searchView.findViewById(androidx.appcompat.R.id.search_edit_frame).setLayoutDirection(View.LAYOUT_DIRECTION_INHERIT);
+        }
+    }
+
+    @Override
+    public void onCreateMenu(@NonNull Menu menu, @NonNull MenuInflater menuInflater) {
+
+    }
+
+    @Override
+    public boolean onMenuItemSelected(@NonNull MenuItem menuItem) {
+        return false;
     }
 
     @Override
@@ -265,7 +278,7 @@ public class ModulesFragment extends BaseFragment implements ModuleUtil.ModuleLi
     }
 
     void installModuleToUser(ModuleUtil.InstalledModule module, UserInfo user) {
-        new BlurBehindDialogBuilder(requireActivity())
+        new BlurBehindDialogBuilder(requireActivity(), R.style.ThemeOverlay_MaterialAlertDialog_Centered_FullWidthButtons)
                 .setTitle(getString(R.string.install_to_user, user.name))
                 .setMessage(getString(R.string.install_to_user_message, module.getAppName(), user.name))
                 .setPositiveButton(android.R.string.ok, (dialog, which) ->
@@ -309,7 +322,8 @@ public class ModulesFragment extends BaseFragment implements ModuleUtil.ModuleLi
             ConfigManager.startActivityAsUserWithFeature(new Intent(ACTION_APPLICATION_DETAILS_SETTINGS, Uri.fromParts("package", selectedModule.packageName, null)), selectedModule.userId);
             return true;
         } else if (itemId == R.id.menu_uninstall) {
-            new BlurBehindDialogBuilder(requireActivity())
+            new BlurBehindDialogBuilder(requireActivity(), R.style.ThemeOverlay_MaterialAlertDialog_FullWidthButtons)
+                    .setIcon(selectedModule.app.loadIcon(pm))
                     .setTitle(selectedModule.getAppName())
                     .setMessage(R.string.module_uninstall_message)
                     .setPositiveButton(android.R.string.ok, (dialog, which) ->
@@ -355,12 +369,12 @@ public class ModulesFragment extends BaseFragment implements ModuleUtil.ModuleLi
 
         private final View.OnAttachStateChangeListener searchViewLocker = new View.OnAttachStateChangeListener() {
             @Override
-            public void onViewAttachedToWindow(View v) {
+            public void onViewAttachedToWindow(@NonNull View v) {
                 binding.recyclerView.setNestedScrollingEnabled(false);
             }
 
             @Override
-            public void onViewDetachedFromWindow(View v) {
+            public void onViewDetachedFromWindow(@NonNull View v) {
                 binding.recyclerView.setNestedScrollingEnabled(true);
             }
         };
@@ -387,8 +401,7 @@ public class ModulesFragment extends BaseFragment implements ModuleUtil.ModuleLi
 
         void attachListeners() {
             var parent = getParentFragment();
-            if (parent instanceof ModulesFragment) {
-                var moduleFragment = (ModulesFragment) parent;
+            if (parent instanceof ModulesFragment moduleFragment) {
                 binding.recyclerView.getBorderViewDelegate().setBorderVisibilityChangedListener((top, oldTop, bottom, oldBottom) -> moduleFragment.binding.appBar.setLifted(!top));
                 moduleFragment.binding.appBar.setLifted(!binding.recyclerView.getBorderViewDelegate().isShowingTopBorder());
                 moduleFragment.searchView.addOnAttachStateChangeListener(searchViewLocker);
@@ -407,8 +420,7 @@ public class ModulesFragment extends BaseFragment implements ModuleUtil.ModuleLi
         void detachListeners() {
             binding.recyclerView.getBorderViewDelegate().setBorderVisibilityChangedListener(null);
             var parent = getParentFragment();
-            if (parent instanceof ModulesFragment) {
-                var moduleFragment = (ModulesFragment) parent;
+            if (parent instanceof ModulesFragment moduleFragment) {
                 moduleFragment.searchView.removeOnAttachStateChangeListener(searchViewLocker);
                 binding.recyclerView.setNestedScrollingEnabled(true);
             }
@@ -505,7 +517,7 @@ public class ModulesFragment extends BaseFragment implements ModuleUtil.ModuleLi
         @NonNull
         @Override
         public ModuleAdapter.ViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
-            return new ModuleAdapter.ViewHolder(ItemModuleBinding.inflate(getLayoutInflater(), parent, false));
+            return new ViewHolder(ItemModuleBinding.inflate(getLayoutInflater(), parent, false));
         }
 
         public boolean isPick() {
@@ -542,7 +554,7 @@ public class ModulesFragment extends BaseFragment implements ModuleUtil.ModuleLi
                 sb.append(getString(R.string.module_empty_description));
             }
             holder.appDescription.setText(sb);
-
+            holder.appDescription.setVisibility(View.VISIBLE);
             sb = new SpannableStringBuilder();
 
             int installXposedVersion = ConfigManager.getXposedApiVersion();
@@ -550,9 +562,11 @@ public class ModulesFragment extends BaseFragment implements ModuleUtil.ModuleLi
             if (item.minVersion == 0) {
                 warningText = getString(R.string.no_min_version_specified);
             } else if (installXposedVersion > 0 && item.minVersion > installXposedVersion) {
-                warningText = String.format(getString(R.string.warning_xposed_min_version), item.minVersion);
+                warningText = getString(R.string.warning_xposed_min_version, item.minVersion);
+            } else if (item.targetVersion > installXposedVersion) {
+                warningText = getString(R.string.warning_target_version_higher, item.targetVersion);
             } else if (item.minVersion < ModuleUtil.MIN_MODULE_VERSION) {
-                warningText = String.format(getString(R.string.warning_min_version_too_low), item.minVersion, ModuleUtil.MIN_MODULE_VERSION);
+                warningText = getString(R.string.warning_min_version_too_low, item.minVersion, ModuleUtil.MIN_MODULE_VERSION);
             } else if (item.isInstalledOnExternalStorage()) {
                 warningText = getString(R.string.warning_installed_on_external_storage);
             }
@@ -594,7 +608,7 @@ public class ModulesFragment extends BaseFragment implements ModuleUtil.ModuleLi
                 holder.root.setAlpha(moduleUtil.isModuleEnabled(item.packageName) ? 1.0f : .5f);
                 holder.itemView.setOnClickListener(v -> {
                     searchView.clearFocus();
-                    getNavController().navigate(ModulesFragmentDirections.actionModulesFragmentToAppListFragment(item.packageName, item.userId));
+                    safeNavigate(ModulesFragmentDirections.actionModulesFragmentToAppListFragment(item.packageName, item.userId));
                 });
                 holder.itemView.setOnLongClickListener(v -> {
                     searchView.clearFocus();
@@ -668,7 +682,7 @@ public class ModulesFragment extends BaseFragment implements ModuleUtil.ModuleLi
 
         public void fullRefresh() {
             runAsync(() -> {
-                setLoaded(false);
+                setLoaded(null, false);
                 moduleUtil.reloadInstalledModules();
                 refresh();
             });
@@ -678,7 +692,7 @@ public class ModulesFragment extends BaseFragment implements ModuleUtil.ModuleLi
             var modules = moduleUtil.getModules();
             if (modules == null) return;
             Comparator<PackageInfo> cmp = AppHelper.getAppListComparator(0, pm);
-            setLoaded(false);
+            setLoaded(null, false);
             var tmpList = new ArrayList<ModuleUtil.InstalledModule>();
             modules.values().parallelStream()
                     .sorted((a, b) -> {
@@ -698,29 +712,30 @@ public class ModulesFragment extends BaseFragment implements ModuleUtil.ModuleLi
                             return 1;
                         }
                     }).forEachOrdered(new Consumer<>() {
-                private final HashSet<String> uniquer = new HashSet<>();
+                        private final HashSet<String> uniquer = new HashSet<>();
 
-                @Override
-                public void accept(ModuleUtil.InstalledModule module) {
-                    if (isPick()) {
-                        if (!uniquer.contains(module.packageName)) {
-                            uniquer.add(module.packageName);
-                            if (module.userId != getUser().id)
+                        @Override
+                        public void accept(ModuleUtil.InstalledModule module) {
+                            if (isPick()) {
+                                if (!uniquer.contains(module.packageName)) {
+                                    uniquer.add(module.packageName);
+                                    if (module.userId != getUser().id)
+                                        tmpList.add(module);
+                                }
+                            } else if (module.userId == getUser().id) {
                                 tmpList.add(module);
+                            }
                         }
-                    } else if (module.userId == getUser().id) {
-                        tmpList.add(module);
-                    }
-                }
-            });
+                    });
             String queryStr = searchView != null ? searchView.getQuery().toString() : "";
             searchList = tmpList;
             runOnUiThread(() -> getFilter().filter(queryStr));
         };
 
         @SuppressLint("NotifyDataSetChanged")
-        private void setLoaded(boolean loaded) {
+        private void setLoaded(List<ModuleUtil.InstalledModule> list, boolean loaded) {
             runOnUiThread(() -> {
+                if (list != null) showList = list;
                 isLoaded = loaded;
                 notifyDataSetChanged();
             });
@@ -731,7 +746,7 @@ public class ModulesFragment extends BaseFragment implements ModuleUtil.ModuleLi
             return isLoaded && moduleUtil.isModulesLoaded();
         }
 
-        class ViewHolder extends RecyclerView.ViewHolder {
+        static class ViewHolder extends RecyclerView.ViewHolder {
             ConstraintLayout root;
             ImageView appIcon;
             TextView appName;
@@ -778,8 +793,7 @@ public class ModulesFragment extends BaseFragment implements ModuleUtil.ModuleLi
             @Override
             protected void publishResults(CharSequence constraint, FilterResults results) {
                 //noinspection unchecked
-                showList = (List<ModuleUtil.InstalledModule>) results.values;
-                setLoaded(true);
+                setLoaded((List<ModuleUtil.InstalledModule>) results.values, true);
             }
         }
     }
