@@ -19,6 +19,8 @@
 
 package org.lsposed.manager.ui.fragment;
 
+import android.annotation.SuppressLint;
+import android.app.Activity;
 import android.app.Dialog;
 import android.content.res.Resources;
 import android.graphics.Color;
@@ -26,9 +28,14 @@ import android.os.Bundle;
 import android.text.Spannable;
 import android.text.SpannableStringBuilder;
 import android.text.TextUtils;
+import android.text.format.Formatter;
 import android.text.style.ClickableSpan;
+import android.text.style.ForegroundColorSpan;
+import android.text.style.RelativeSizeSpan;
 import android.util.Log;
 import android.view.LayoutInflater;
+import android.view.Menu;
+import android.view.MenuInflater;
 import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
@@ -37,11 +44,13 @@ import android.webkit.WebResourceResponse;
 import android.webkit.WebSettings;
 import android.webkit.WebView;
 import android.webkit.WebViewClient;
+import android.widget.ArrayAdapter;
 import android.widget.ScrollView;
 import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+import androidx.core.view.MenuProvider;
 import androidx.fragment.app.DialogFragment;
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentManager;
@@ -76,6 +85,10 @@ import org.lsposed.manager.util.chrome.CustomTabsURLSpan;
 
 import java.io.ByteArrayInputStream;
 import java.nio.charset.StandardCharsets;
+import java.time.Instant;
+import java.time.ZoneId;
+import java.time.format.DateTimeFormatter;
+import java.time.format.FormatStyle;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.ListIterator;
@@ -90,7 +103,7 @@ import rikka.material.app.LocaleDelegate;
 import rikka.recyclerview.RecyclerViewKt;
 import rikka.widget.borderview.BorderView;
 
-public class RepoItemFragment extends BaseFragment implements RepoLoader.RepoListener {
+public class RepoItemFragment extends BaseFragment implements RepoLoader.RepoListener, MenuProvider {
     FragmentPagerBinding binding;
     OnlineModule module;
     private ReleaseAdapter releaseAdapter;
@@ -133,17 +146,19 @@ public class RepoItemFragment extends BaseFragment implements RepoLoader.RepoLis
 
         String modulePackageName = getArguments() == null ? null : getArguments().getString("modulePackageName");
         module = RepoLoader.getInstance().getOnlineModule(modulePackageName);
-        if (module == null)
-            getNavController().navigate(R.id.action_repo_item_fragment_to_repo_fragment);
+        if (module == null) {
+            if (!safeNavigate(R.id.action_repo_item_fragment_to_repo_fragment)) {
+                safeNavigate(R.id.repo_nav);
+            }
+        }
     }
 
-    private void renderGithubMarkdown(WebView view, String text) {
+    private void renderGithubMarkdown(WebView view, @Nullable String text) {
         try {
             view.setBackgroundColor(Color.TRANSPARENT);
             var setting = view.getSettings();
             setting.setOffscreenPreRaster(true);
             setting.setDomStorageEnabled(true);
-            setting.setAppCacheEnabled(true);
             setting.setMixedContentMode(WebSettings.MIXED_CONTENT_ALWAYS_ALLOW);
             setting.setAllowContentAccess(false);
             setting.setAllowFileAccessFromFileURLs(true);
@@ -157,6 +172,9 @@ public class RepoItemFragment extends BaseFragment implements RepoLoader.RepoLis
                 direction = "rtl";
             } else {
                 direction = "ltr";
+            }
+            if (text == null) {
+                text = "<center>" + App.getInstance().getString(R.string.list_empty) + "</center>";
             }
             if (ResourceUtils.isNightMode(getResources().getConfiguration())) {
                 body = App.HTML_TEMPLATE_DARK.get().replace("@dir@", direction).replace("@body@", text);
@@ -210,12 +228,18 @@ public class RepoItemFragment extends BaseFragment implements RepoLoader.RepoLis
     }
 
     @Override
-    public boolean onOptionsItemSelected(@NonNull MenuItem item) {
+    public void onCreateMenu(@NonNull Menu menu, @NonNull MenuInflater menuInflater) {
+
+    }
+
+    @Override
+    public boolean onMenuItemSelected(@NonNull MenuItem item) {
         int id = item.getItemId();
         if (id == R.id.menu_open_in_browser) {
             NavUtil.startURL(requireActivity(), "https://modules.lsposed.org/module/" + module.getName());
+            return true;
         }
-        return super.onOptionsItemSelected(item);
+        return false;
     }
 
     @Override
@@ -228,10 +252,11 @@ public class RepoItemFragment extends BaseFragment implements RepoLoader.RepoLis
     @Override
     public void onModuleReleasesLoaded(OnlineModule module) {
         this.module = module;
+        var repoLoader = RepoLoader.getInstance();
         if (releaseAdapter != null) {
             runAsync(releaseAdapter::loadItems);
         }
-        if (module.getReleases().size() == 1) {
+        if ((repoLoader.getReleases(module.getName()) != null ? repoLoader.getReleases(module.getName()).size() : 1) == 1) {
             showHint(R.string.module_release_no_more, true);
         }
     }
@@ -275,13 +300,16 @@ public class RepoItemFragment extends BaseFragment implements RepoLoader.RepoLis
                 holder.title.setText(R.string.module_information_homepage);
                 holder.description.setText(module.getHomepageUrl());
             } else if (position == collaboratorsRow) {
-                holder.title.setText(R.string.module_information_collaborators);
                 List<Collaborator> collaborators = module.getCollaborators();
+                if (collaborators == null) return;
+                holder.title.setText(R.string.module_information_collaborators);
                 SpannableStringBuilder sb = new SpannableStringBuilder();
                 ListIterator<Collaborator> iterator = collaborators.listIterator();
                 while (iterator.hasNext()) {
                     Collaborator collaborator = iterator.next();
-                    String name = collaborator.getName() == null ? collaborator.getLogin() : collaborator.getName();
+                    var collaboratorLogin = collaborator.getLogin();
+                    if (collaboratorLogin == null) continue;
+                    String name = collaborator.getName() == null ? collaboratorLogin : collaborator.getName();
                     sb.append(name);
                     CustomTabsURLSpan span = new CustomTabsURLSpan(requireActivity(), String.format("https://github.com/%s", collaborator.getLogin()));
                     sb.setSpan(span, sb.length() - name.length(), sb.length(), Spannable.SPAN_EXCLUSIVE_EXCLUSIVE);
@@ -333,16 +361,33 @@ public class RepoItemFragment extends BaseFragment implements RepoLoader.RepoLis
         public Dialog onCreateDialog(@Nullable Bundle savedInstanceState) {
             var args = getArguments();
             if (args == null) throw new IllegalArgumentException();
-            return new BlurBehindDialogBuilder(requireActivity())
-                    .setItems(args.getCharSequenceArray("names"), (dialog, which) -> NavUtil.startURL(requireActivity(), args.getStringArrayList("urls").get(which)))
+            return new BlurBehindDialogBuilder(requireActivity(), R.style.ThemeOverlay_MaterialAlertDialog_Centered_FullWidthButtons)
+                    .setTitle(R.string.module_release_view_assets)
+                    .setPositiveButton(android.R.string.cancel, null)
+                    .setAdapter(new ArrayAdapter<>(requireActivity(), R.layout.dialog_item, args.getCharSequenceArray("names")),
+                            (dialog, which) -> NavUtil.startURL(requireActivity(), args.getStringArrayList("urls").get(which)))
                     .create();
         }
 
-        static void create(FragmentManager fm, String[] names, ArrayList<String> urls) {
+        static void create(Activity activity, FragmentManager fm, List<ReleaseAsset> assets) {
             var f = new DownloadDialog();
             var bundle = new Bundle();
-            bundle.putStringArray("names", names);
-            bundle.putStringArrayList("urls", urls);
+
+            var displayNames = new CharSequence[assets.size()];
+            for (int i = 0; i < assets.size(); i++) {
+                var sb = new SpannableStringBuilder(assets.get(i).getName());
+                var count = assets.get(i).getDownloadCount();
+                var countStr = activity.getResources().getQuantityString(R.plurals.module_release_assets_download_count, count, count);
+                var sizeStr = Formatter.formatShortFileSize(activity, assets.get(i).getSize());
+                sb.append('\n').append(sizeStr).append('/').append(countStr);
+                final ForegroundColorSpan foregroundColorSpan = new ForegroundColorSpan(ResourceUtils.resolveColor(activity.getTheme(), android.R.attr.textColorSecondary));
+                final RelativeSizeSpan relativeSizeSpan = new RelativeSizeSpan(0.8f);
+                sb.setSpan(foregroundColorSpan, sb.length() - sizeStr.length() - countStr.length() - 1, sb.length(), Spannable.SPAN_EXCLUSIVE_EXCLUSIVE);
+                sb.setSpan(relativeSizeSpan, sb.length() - sizeStr.length() - countStr.length() - 1, sb.length(), Spannable.SPAN_EXCLUSIVE_EXCLUSIVE);
+                displayNames[i] = sb;
+            }
+            bundle.putCharSequenceArray("names", displayNames);
+            bundle.putStringArrayList("urls", assets.stream().map(ReleaseAsset::getDownloadUrl).collect(Collectors.toCollection(ArrayList::new)));
             f.setArguments(bundle);
             f.show(fm, "download");
         }
@@ -356,23 +401,29 @@ public class RepoItemFragment extends BaseFragment implements RepoLoader.RepoLis
             runAsync(this::loadItems);
         }
 
+        @SuppressLint("NotifyDataSetChanged")
         public void loadItems() {
             var channels = resources.getStringArray(R.array.update_channel_values);
             var channel = App.getPreferences().getString("update_channel", channels[0]);
-            var releases = module.getReleases();
+            var releases = RepoLoader.getInstance().getReleases(module.getName());
+            if (releases == null) releases = module.getReleases();
+            List<Release> tmpList;
             if (channel.equals(channels[0])) {
-                this.items = releases.parallelStream().filter(t -> {
-                    if (t.getIsPrerelease()) return false;
-                    var name = t.getName().toLowerCase(LocaleDelegate.getDefaultLocale());
-                    return !name.startsWith("snapshot") && !name.startsWith("nightly");
-                }).collect(Collectors.toList());
+                tmpList = releases != null ? releases.parallelStream().filter(t -> {
+                    if (Boolean.TRUE.equals(t.getIsPrerelease())) return false;
+                    var name = t.getName() != null ? t.getName().toLowerCase(LocaleDelegate.getDefaultLocale()) : null;
+                    return !(name != null && name.startsWith("snapshot")) && !(name != null && name.startsWith("nightly"));
+                }).collect(Collectors.toList()) : null;
             } else if (channel.equals(channels[1])) {
-                this.items = releases.parallelStream().filter(t -> {
-                    var name = t.getName().toLowerCase(LocaleDelegate.getDefaultLocale());
-                    return !name.startsWith("snapshot") && !name.startsWith("nightly");
-                }).collect(Collectors.toList());
-            } else this.items = releases;
-            runOnUiThread(this::notifyDataSetChanged);
+                tmpList = releases != null ? releases.parallelStream().filter(t -> {
+                    var name = t.getName() != null ? t.getName().toLowerCase(LocaleDelegate.getDefaultLocale()) : null;
+                    return !(name != null && name.startsWith("snapshot")) && !(name != null && name.startsWith("nightly"));
+                }).collect(Collectors.toList()) : null;
+            } else tmpList = releases;
+            runOnUiThread(() -> {
+                items = tmpList;
+                notifyDataSetChanged();
+            });
         }
 
         @NonNull
@@ -400,15 +451,15 @@ public class RepoItemFragment extends BaseFragment implements RepoLoader.RepoLis
             } else {
                 Release release = items.get(position);
                 holder.title.setText(release.getName());
+                var instant = Instant.parse(release.getPublishedAt());
+                var formatter = DateTimeFormatter.ofLocalizedDateTime(FormatStyle.SHORT)
+                        .withLocale(App.getLocale()).withZone(ZoneId.systemDefault());
+                holder.publishedTime.setText(String.format(getString(R.string.module_repo_published_time), formatter.format(instant)));
                 renderGithubMarkdown(holder.description, release.getDescriptionHTML());
                 holder.openInBrowser.setOnClickListener(v -> NavUtil.startURL(requireActivity(), release.getUrl()));
                 List<ReleaseAsset> assets = release.getReleaseAssets();
                 if (assets != null && !assets.isEmpty()) {
-                    holder.viewAssets.setOnClickListener(v -> {
-                        ArrayList<String> names = new ArrayList<>();
-                        assets.forEach(releaseAsset -> names.add(releaseAsset.getName()));
-                        DownloadDialog.create(getChildFragmentManager(), names.toArray(new String[0]), assets.stream().map(ReleaseAsset::getDownloadUrl).collect(Collectors.toCollection(ArrayList::new)));
-                    });
+                    holder.viewAssets.setOnClickListener(v -> DownloadDialog.create(requireActivity(), getParentFragmentManager(), assets));
                 } else {
                     holder.viewAssets.setVisibility(View.GONE);
                 }
@@ -432,6 +483,7 @@ public class RepoItemFragment extends BaseFragment implements RepoLoader.RepoLis
 
         class ViewHolder extends RecyclerView.ViewHolder {
             TextView title;
+            TextView publishedTime;
             WebView description;
             MaterialButton openInBrowser;
             MaterialButton viewAssets;
@@ -446,6 +498,7 @@ public class RepoItemFragment extends BaseFragment implements RepoLoader.RepoLis
             public ReleaseViewHolder(ItemRepoReleaseBinding binding) {
                 super(binding.getRoot());
                 title = binding.title;
+                publishedTime = binding.publishedTime;
                 description = binding.description;
                 openInBrowser = binding.openInBrowser;
                 viewAssets = binding.viewAssets;
@@ -555,7 +608,9 @@ public class RepoItemFragment extends BaseFragment implements RepoLoader.RepoLis
         public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
             var parent = getParentFragment();
             if (!(parent instanceof RepoItemFragment)) {
-                getNavController().navigate(R.id.action_repo_item_fragment_to_repo_fragment);
+                if (!safeNavigate(R.id.action_repo_item_fragment_to_repo_fragment)) {
+                    safeNavigate(R.id.repo_nav);
+                }
                 return null;
             }
             var repoItemFragment = (RepoItemFragment) parent;
@@ -584,7 +639,9 @@ public class RepoItemFragment extends BaseFragment implements RepoLoader.RepoLis
             var arguments = getArguments();
             var parent = getParentFragment();
             if (arguments == null || !(parent instanceof RepoItemFragment)) {
-                getNavController().navigate(R.id.action_repo_item_fragment_to_repo_fragment);
+                if (!safeNavigate(R.id.action_repo_item_fragment_to_repo_fragment)) {
+                    safeNavigate(R.id.repo_nav);
+                }
                 return null;
             }
             var repoItemFragment = (RepoItemFragment) parent;
