@@ -19,11 +19,16 @@
 
 package org.lsposed.lspd.service;
 
+import static android.content.Intent.EXTRA_UID;
+import static org.lsposed.lspd.service.LSPNotificationManager.SCOPE_CHANNEL_ID;
+import static org.lsposed.lspd.service.LSPNotificationManager.UPDATED_CHANNEL_ID;
 import static org.lsposed.lspd.service.PackageService.PER_USER_RANGE;
 import static org.lsposed.lspd.service.ServiceManager.TAG;
 import static org.lsposed.lspd.service.ServiceManager.getExecutorService;
 
 import android.app.IApplicationThread;
+import android.app.IUidObserver;
+import android.content.Context;
 import android.content.IIntentReceiver;
 import android.content.Intent;
 import android.content.IntentFilter;
@@ -31,13 +36,25 @@ import android.content.pm.ApplicationInfo;
 import android.content.pm.PackageManager;
 import android.net.Uri;
 import android.os.Binder;
+import android.os.Build;
 import android.os.Bundle;
 import android.os.IBinder;
+import android.os.RemoteException;
+import android.provider.Telephony;
+import android.telephony.TelephonyManager;
 import android.util.Log;
 
 import org.lsposed.daemon.BuildConfig;
 
+import java.io.IOException;
 import java.util.Arrays;
+import java.util.List;
+import java.util.Objects;
+import java.util.function.Consumer;
+import java.util.zip.ZipFile;
+
+import hidden.HiddenApiBridge;
+import io.github.libxposed.service.IXposedScopeCallback;
 
 public class LSPosedService extends ILSPosedService.Stub {
     private static final int AID_NOBODY = 9999;
@@ -46,6 +63,25 @@ public class LSPosedService extends ILSPosedService.Stub {
     public static final String ACTION_USER_REMOVED = "android.intent.action.USER_REMOVED";
     private static final String EXTRA_USER_HANDLE = "android.intent.extra.user_handle";
     private static final String EXTRA_REMOVED_FOR_ALL_USERS = "android.intent.extra.REMOVED_FOR_ALL_USERS";
+    private static boolean bootCompleted = false;
+    private IBinder appThread = null;
+
+    private static boolean isModernModules(ApplicationInfo info) {
+        String[] apks;
+        if (info.splitSourceDirs != null) {
+            apks = Arrays.copyOf(info.splitSourceDirs, info.splitSourceDirs.length + 1);
+            apks[info.splitSourceDirs.length] = info.sourceDir;
+        } else apks = new String[]{info.sourceDir};
+        for (var apk : apks) {
+            try (var zip = new ZipFile(apk)) {
+                if (zip.getEntry("META-INF/xposed/java_init.list") != null) {
+                    return true;
+                }
+            } catch (IOException ignored) {
+            }
+        }
+        return false;
+    }
 
     @Override
     public ILSPApplicationService requestApplicationService(int uid, int pid, String processName, IBinder heartBeat) {
@@ -57,13 +93,12 @@ public class LSPosedService extends ILSPosedService.Stub {
             Log.d(TAG, "Skipped duplicated request for uid " + uid + " pid " + pid);
             return null;
         }
-        if (!ServiceManager.getManagerService().shouldStartManager(pid, uid, processName) &&
-                ConfigManager.getInstance().shouldSkipProcess(new ConfigManager.ProcessScope(processName, uid))) {
+        if (!ServiceManager.getManagerService().shouldStartManager(pid, uid, processName) && ConfigManager.getInstance().shouldSkipProcess(new ConfigManager.ProcessScope(processName, uid))) {
             Log.d(TAG, "Skipped " + processName + "/" + uid);
             return null;
         }
         Log.d(TAG, "returned service");
-        return ServiceManager.requestApplicationService(uid, pid, heartBeat);
+        return ServiceManager.requestApplicationService(uid, pid, processName, heartBeat);
     }
 
     /**
@@ -77,16 +112,18 @@ public class LSPosedService extends ILSPosedService.Stub {
      * However, PACKAGE_REMOVED will be triggered by `pm hide`, so we use UID_REMOVED instead.
      */
 
-    synchronized public void dispatchPackageChanged(Intent intent) {
+    private void dispatchPackageChanged(Intent intent) {
         if (intent == null) return;
-        int uid = intent.getIntExtra(Intent.EXTRA_UID, AID_NOBODY);
+        int uid = intent.getIntExtra(EXTRA_UID, AID_NOBODY);
         if (uid == AID_NOBODY || uid <= 0) return;
         int userId = intent.getIntExtra("android.intent.extra.user_handle", USER_NULL);
         var intentAction = intent.getAction();
+        if (intentAction == null) return;
         var allUsers = intent.getBooleanExtra(EXTRA_REMOVED_FOR_ALL_USERS, false);
         if (userId == USER_NULL) userId = uid % PER_USER_RANGE;
         Uri uri = intent.getData();
-        String moduleName = (uri != null) ? uri.getSchemeSpecificPart() : ConfigManager.getInstance().getModule(uid);
+        var module = ConfigManager.getInstance().getModule(uid);
+        String moduleName = (uri != null) ? uri.getSchemeSpecificPart() : (module != null) ? module.packageName : null;
 
         ApplicationInfo applicationInfo = null;
         if (moduleName != null) {
@@ -96,70 +133,64 @@ public class LSPosedService extends ILSPosedService.Stub {
             }
         }
 
-        boolean isXposedModule = applicationInfo != null &&
-                applicationInfo.metaData != null &&
-                applicationInfo.metaData.containsKey("xposedminversion");
+        boolean isXposedModule = applicationInfo != null && ((applicationInfo.metaData != null && applicationInfo.metaData.containsKey("xposedminversion")) || isModernModules(applicationInfo));
 
         switch (intentAction) {
-            case Intent.ACTION_PACKAGE_FULLY_REMOVED: {
+            case Intent.ACTION_PACKAGE_FULLY_REMOVED -> {
                 // for module, remove module
                 // because we only care about when the apk is gone
-                if (moduleName != null && allUsers)
-                    if (ConfigManager.getInstance().removeModule(moduleName)) {
+                if (moduleName != null) {
+                    if (allUsers && ConfigManager.getInstance().removeModule(moduleName)) {
                         isXposedModule = true;
                         broadcastAndShowNotification(moduleName, userId, intent, true);
                     }
-                break;
+                    LSPNotificationManager.cancelNotification(UPDATED_CHANNEL_ID, moduleName, userId);
+                }
             }
-            case Intent.ACTION_PACKAGE_ADDED:
-            case Intent.ACTION_PACKAGE_CHANGED: {
+            case Intent.ACTION_PACKAGE_REMOVED -> {
+                if (moduleName != null) {
+                    LSPNotificationManager.cancelNotification(UPDATED_CHANNEL_ID, moduleName, userId);
+                }
+            }
+            case Intent.ACTION_PACKAGE_ADDED, Intent.ACTION_PACKAGE_CHANGED -> {
                 // make sure that the change is for the complete package, not only a
                 // component
                 String[] components = intent.getStringArrayExtra(Intent.EXTRA_CHANGED_COMPONENT_NAME_LIST);
                 if (components != null && !Arrays.stream(components).reduce(false, (p, c) -> p || c.equals(moduleName), Boolean::logicalOr)) {
                     return;
                 }
-                broadcastAndShowNotification(moduleName, userId, intent, isXposedModule);
                 if (isXposedModule) {
                     // When installing a new Xposed module, we update the apk path to mark it as a
                     // module to send a broadcast when modules that have not been activated are
                     // uninstalled.
-                    ConfigManager.getInstance().updateModuleApkPath(moduleName, ConfigManager.getInstance().getModuleApkPath(applicationInfo), true);
-                    // when package is changed, we may need to update cache (module cache or process cache)
-                    ConfigManager.getInstance().updateCache();
+                    // If cache not updated, assume it's not xposed module
+                    isXposedModule = ConfigManager.getInstance().updateModuleApkPath(moduleName, ConfigManager.getInstance().getModuleApkPath(applicationInfo), false);
                 } else if (ConfigManager.getInstance().isUidHooked(uid)) {
-                    // it will automatically remove obsolete app from database
+                    // it will auto update obsolete scope from database
                     ConfigManager.getInstance().updateAppCache();
                 }
-                break;
+                broadcastAndShowNotification(moduleName, userId, intent, isXposedModule);
             }
-            case Intent.ACTION_UID_REMOVED: {
+            case Intent.ACTION_UID_REMOVED -> {
                 // when a package is removed (rather than hide) for a single user
                 // (apk may still be there because of multi-user)
                 broadcastAndShowNotification(moduleName, userId, intent, isXposedModule);
                 if (isXposedModule) {
-                    // it will automatically remove obsolete scope from database
+                    // it will auto remove obsolete app and scope from database
                     ConfigManager.getInstance().updateCache();
                 } else if (ConfigManager.getInstance().isUidHooked(uid)) {
-                    // it will automatically remove obsolete app from database
+                    // it will auto remove obsolete scope from database
                     ConfigManager.getInstance().updateAppCache();
                 }
-                break;
             }
         }
-        boolean removed = Intent.ACTION_PACKAGE_FULLY_REMOVED.equals(intentAction) ||
-                Intent.ACTION_UID_REMOVED.equals(intentAction);
+        boolean removed = Intent.ACTION_PACKAGE_FULLY_REMOVED.equals(intentAction) || Intent.ACTION_UID_REMOVED.equals(intentAction);
 
         Log.d(TAG, "Package changed: uid=" + uid + " userId=" + userId + " action=" + intentAction + " isXposedModule=" + isXposedModule + " isAllUsers=" + allUsers);
 
         if (BuildConfig.DEFAULT_MANAGER_PACKAGE_NAME.equals(moduleName) && userId == 0) {
             Log.d(TAG, "Manager updated");
-            try {
-                ConfigManager.getInstance().updateManager(removed);
-                LSPManagerService.createOrUpdateShortcut(false);
-            } catch (Throwable e) {
-                Log.e(TAG, Log.getStackTraceString(e));
-            }
+            ConfigManager.getInstance().updateManager(removed);
         }
     }
 
@@ -174,228 +205,266 @@ public class LSPosedService extends ILSPosedService.Stub {
         if (isXposedModule) {
             var enabledModules = ConfigManager.getInstance().enabledModules();
             var scope = ConfigManager.getInstance().getModuleScope(packageName);
-            boolean systemModule = scope != null &&
-                    scope.parallelStream().anyMatch(app -> app.packageName.equals("android"));
+            boolean systemModule = scope != null && scope.parallelStream().anyMatch(app -> app.packageName.equals("system"));
             boolean enabled = Arrays.asList(enabledModules).contains(packageName);
             if (!(Intent.ACTION_UID_REMOVED.equals(action) || Intent.ACTION_PACKAGE_FULLY_REMOVED.equals(action) || allUsers))
-                LSPManagerService.showNotification(packageName, userId, enabled, systemModule);
+                LSPNotificationManager.notifyModuleUpdated(packageName, userId, enabled, systemModule);
         }
     }
 
-    synchronized public void dispatchUserChanged(Intent intent) {
+    private void dispatchUserChanged(Intent intent) {
         if (intent == null) return;
         int uid = intent.getIntExtra(EXTRA_USER_HANDLE, AID_NOBODY);
         if (uid == AID_NOBODY || uid <= 0) return;
+        LSPManagerService.broadcastIntent(intent);
+    }
+
+    private void dispatchBootCompleted(Intent intent) {
+        bootCompleted = true;
         try {
-            LSPManagerService.broadcastIntent(intent);
-        } catch (Throwable e) {
-            Log.e(TAG, "dispatch user info changed", e);
+            var am = ActivityManagerService.getActivityManager();
+            if (am != null) am.setActivityController(null, false);
+        } catch (RemoteException e) {
+            Log.e(TAG, "setActivityController", e);
+        }
+        var configManager = ConfigManager.getInstance();
+        if (configManager.enableStatusNotification()) {
+            LSPNotificationManager.notifyStatusNotification();
+        } else {
+            LSPNotificationManager.cancelStatusNotification();
         }
     }
 
-    synchronized public void dispatchUserUnlocked(Intent intent) {
-        try {
-            LSPManagerService.createOrUpdateShortcut(false);
-        } catch (Throwable e) {
-            Log.e(TAG, "dispatch user unlocked", e);
+    private void dispatchConfigurationChanged(Intent intent) {
+        if (!bootCompleted) return;
+        ConfigFileManager.reloadConfiguration();
+        var configManager = ConfigManager.getInstance();
+        if (configManager.enableStatusNotification()) {
+            LSPNotificationManager.notifyStatusNotification();
+        } else {
+            LSPNotificationManager.cancelStatusNotification();
         }
     }
 
-    synchronized public void dispatchConfigurationChanged(Intent intent) {
-        try {
-            ConfigFileManager.reloadConfiguration();
-            LSPManagerService.createOrUpdateShortcut(false, false);
-        } catch (Throwable e) {
-            Log.e(TAG, "dispatch configuration changed", e);
-        }
+    private void dispatchSecretCodeReceive(Intent i) {
+        LSPManagerService.openManager(null);
     }
 
-    synchronized public void dispatchSecretCodeReceive() {
-        Intent intent = LSPManagerService.getManagerIntent();
+    private void dispatchOpenManager(Intent intent) {
+        LSPManagerService.openManager(intent.getData());
+    }
+
+    private void dispatchModuleScope(Intent intent) {
+        Log.d(TAG, "dispatchModuleScope: " + intent);
+        var data = intent.getData();
+        var extras = intent.getExtras();
+        if (extras == null || data == null) return;
+        var callback = extras.getBinder("callback");
+        if (callback == null || !callback.isBinderAlive()) return;
+        var authority = data.getEncodedAuthority();
+        if (authority == null) return;
+        var s = authority.split(":", 2);
+        if (s.length != 2) return;
+        var packageName = s[0];
+        int userId;
         try {
-            var userInfo = ActivityManagerService.getCurrentUser();
-            if (userInfo != null) {
-                var userId = userInfo.id;
-                if (userId == 0) {
-                    ActivityManagerService.startActivityAsUserWithFeature("android", null,
-                            intent, intent.getType(), null, null, 0, 0, null, null, userId);
-                    LSPManagerService.createOrUpdateShortcut(false);
+            userId = Integer.parseInt(s[1]);
+        } catch (NumberFormatException e) {
+            return;
+        }
+        var scopePackageName = data.getPath();
+        if (scopePackageName == null) return;
+        scopePackageName = scopePackageName.substring(1);
+        var action = data.getQueryParameter("action");
+        if (action == null) return;
+
+        var iCallback = IXposedScopeCallback.Stub.asInterface(callback);
+        try {
+            var applicationInfo = PackageService.getApplicationInfo(scopePackageName, 0, userId);
+            if (applicationInfo == null) {
+                iCallback.onScopeRequestFailed(scopePackageName, "Package not found");
+                return;
+            }
+
+            switch (action) {
+                case "approve" -> {
+                    ConfigManager.getInstance().setModuleScope(packageName, scopePackageName, userId);
+                    iCallback.onScopeRequestApproved(scopePackageName);
+                }
+                case "deny" -> iCallback.onScopeRequestDenied(scopePackageName);
+                case "delete" -> iCallback.onScopeRequestTimeout(scopePackageName);
+                case "block" -> {
+                    ConfigManager.getInstance().blockScopeRequest(packageName);
+                    iCallback.onScopeRequestDenied(scopePackageName);
                 }
             }
-        } catch (Throwable e) {
-            Log.e(TAG, "dispatch secret code received", e);
+            Log.i(TAG, action + " scope " + scopePackageName + " for " + packageName + " in user " + userId);
+        } catch (RemoteException e) {
+            try {
+                iCallback.onScopeRequestFailed(scopePackageName, e.getMessage());
+            } catch (RemoteException ignored) {
+                // callback died
+            }
         }
+        LSPNotificationManager.cancelNotification(SCOPE_CHANNEL_ID, packageName, userId);
+    }
+
+    private void registerReceiver(List<IntentFilter> filters, String requiredPermission, int userId, Consumer<Intent> task, int flag) {
+        var receiver = new IIntentReceiver.Stub() {
+            @Override
+            public void performReceive(Intent intent, int resultCode, String data, Bundle extras, boolean ordered, boolean sticky, int sendingUser) {
+                getExecutorService().submit(() -> {
+                    try {
+                        task.accept(intent);
+                    } catch (Throwable t) {
+                        Log.e(TAG, "performReceive: ", t);
+                    }
+                });
+                if (!ordered && !Objects.equals(intent.getAction(), Intent.ACTION_LOCKED_BOOT_COMPLETED))
+                    return;
+                try {
+                    ActivityManagerService.finishReceiver(this, appThread, resultCode, data, extras, false, intent.getFlags());
+                } catch (RemoteException e) {
+                    Log.e(TAG, "finish receiver", e);
+                }
+            }
+        };
+        try {
+            for (var filter : filters) {
+                ActivityManagerService.registerReceiver("android", null, receiver, filter, requiredPermission, userId, flag);
+            }
+        } catch (RemoteException e) {
+            Log.e(TAG, "register receiver", e);
+        }
+    }
+
+    private void registerReceiver(List<IntentFilter> filters, int userId, Consumer<Intent> task) {
+        //noinspection InlinedApi
+        registerReceiver(filters, "android.permission.BRICK", userId, task, Context.RECEIVER_NOT_EXPORTED);
     }
 
     private void registerPackageReceiver() {
-        try {
-            IntentFilter packageFilter = new IntentFilter();
-            packageFilter.addAction(Intent.ACTION_PACKAGE_ADDED);
-            packageFilter.addAction(Intent.ACTION_PACKAGE_CHANGED);
-            packageFilter.addAction(Intent.ACTION_PACKAGE_FULLY_REMOVED);
-            packageFilter.addDataScheme("package");
+        var packageFilter = new IntentFilter();
+        packageFilter.addAction(Intent.ACTION_PACKAGE_ADDED);
+        packageFilter.addAction(Intent.ACTION_PACKAGE_CHANGED);
+        packageFilter.addAction(Intent.ACTION_PACKAGE_FULLY_REMOVED);
+        packageFilter.addDataScheme("package");
 
-            IntentFilter uidFilter = new IntentFilter();
-            uidFilter.addAction(Intent.ACTION_UID_REMOVED);
+        var uidFilter = new IntentFilter(Intent.ACTION_UID_REMOVED);
 
-            var receiver = new IIntentReceiver.Stub() {
-                @Override
-                public void performReceive(Intent intent, int resultCode, String data, Bundle extras, boolean ordered, boolean sticky, int sendingUser) {
-                    getExecutorService().submit(() -> dispatchPackageChanged(intent));
-                    try {
-                        ActivityManagerService.finishReceiver(this, resultCode, data, extras, false, intent.getFlags());
-                    } catch (Throwable e) {
-                        Log.e(TAG, "finish receiver", e);
-                    }
-                }
-            };
-
-            ActivityManagerService.registerReceiver("android", null, receiver, packageFilter, null, -1, 0);
-            ActivityManagerService.registerReceiver("android", null, receiver, uidFilter, null, -1, 0);
-        } catch (Throwable e) {
-            Log.e(TAG, "register package receiver", e);
-        }
+        registerReceiver(List.of(packageFilter, uidFilter), -1, this::dispatchPackageChanged);
         Log.d(TAG, "registered package receiver");
     }
 
-    private void registerUnlockReceiver() {
-        try {
-            IntentFilter intentFilter = new IntentFilter();
-            intentFilter.addAction(Intent.ACTION_USER_UNLOCKED);
-
-            ActivityManagerService.registerReceiver("android", null, new IIntentReceiver.Stub() {
-                @Override
-                public void performReceive(Intent intent, int resultCode, String data, Bundle extras, boolean ordered, boolean sticky, int sendingUser) {
-                    getExecutorService().submit(() -> dispatchUserUnlocked(intent));
-                    try {
-                        ActivityManagerService.finishReceiver(this, resultCode, data, extras, false, intent.getFlags());
-                    } catch (Throwable e) {
-                        Log.e(TAG, "finish receiver", e);
-                    }
-                }
-            }, intentFilter, null, 0, 0);
-        } catch (Throwable e) {
-            Log.e(TAG, "register unlock receiver", e);
-        }
-        Log.d(TAG, "registered unlock receiver");
-    }
-
     private void registerConfigurationReceiver() {
-        try {
-            IntentFilter intentFilter = new IntentFilter();
-            intentFilter.addAction(Intent.ACTION_CONFIGURATION_CHANGED);
+        var intentFilter = new IntentFilter(Intent.ACTION_CONFIGURATION_CHANGED);
 
-            ActivityManagerService.registerReceiver("android", null, new IIntentReceiver.Stub() {
-                @Override
-                public void performReceive(Intent intent, int resultCode, String data, Bundle extras, boolean ordered, boolean sticky, int sendingUser) {
-                    getExecutorService().submit(() -> dispatchConfigurationChanged(intent));
-                    try {
-                        ActivityManagerService.finishReceiver(this, resultCode, data, extras, false, intent.getFlags());
-                    } catch (Throwable e) {
-                        Log.e(TAG, "finish receiver", e);
-                    }
-                }
-            }, intentFilter, null, 0, 0);
-        } catch (Throwable e) {
-            Log.e(TAG, "register configuration receiver", e);
-        }
+        registerReceiver(List.of(intentFilter), 0, this::dispatchConfigurationChanged);
         Log.d(TAG, "registered configuration receiver");
     }
 
     private void registerSecretCodeReceiver() {
-        try {
-            IntentFilter intentFilter = new IntentFilter();
-            intentFilter.addAction("android.provider.Telephony.SECRET_CODE");
-            intentFilter.addDataAuthority("5776733", null);
-            intentFilter.addDataScheme("android_secret_code");
-
-            ActivityManagerService.registerReceiver("android", null, new IIntentReceiver.Stub() {
-                @Override
-                public void performReceive(Intent intent, int resultCode, String data, Bundle extras, boolean ordered, boolean sticky, int sendingUser) {
-                    getExecutorService().submit(() -> dispatchSecretCodeReceive());
-                    try {
-                        ActivityManagerService.finishReceiver(this, resultCode, data, extras, false, intent.getFlags());
-                    } catch (Throwable e) {
-                        Log.e(TAG, "finish receiver", e);
-                    }
-                }
-            }, intentFilter, null, 0, 0);
-        } catch (Throwable e) {
-            Log.e(TAG, "register secret code receiver", e);
+        IntentFilter intentFilter = new IntentFilter();
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
+            intentFilter.addAction(TelephonyManager.ACTION_SECRET_CODE);
+        } else {
+            // noinspection InlinedApi
+            intentFilter.addAction(Telephony.Sms.Intents.SECRET_CODE_ACTION);
         }
+        intentFilter.addDataAuthority("5776733", null);
+        intentFilter.addDataScheme("android_secret_code");
+
+        //noinspection InlinedApi
+        registerReceiver(List.of(intentFilter), "android.permission.CONTROL_INCALL_EXPERIENCE",
+                0, this::dispatchSecretCodeReceive, Context.RECEIVER_EXPORTED);
         Log.d(TAG, "registered secret code receiver");
     }
 
     private void registerBootCompleteReceiver() {
-        try {
-            IntentFilter intentFilter = new IntentFilter();
-            intentFilter.addAction(Intent.ACTION_LOCKED_BOOT_COMPLETED);
-
-            ActivityManagerService.registerReceiver("android", null, new IIntentReceiver.Stub() {
-                @Override
-                public void performReceive(Intent intent, int resultCode, String data, Bundle extras, boolean ordered, boolean sticky, int sendingUser) {
-                    getExecutorService().submit(() -> {
-                        try {
-                            var am = ActivityManagerService.getActivityManager();
-                            if (am != null) am.setActivityController(null, false);
-                        } catch (Throwable e) {
-                            Log.e(TAG, "setActivityController", e);
-                        }
-                    });
-                    try {
-                        ActivityManagerService.finishReceiver(this, resultCode, data, extras, false, intent.getFlags());
-                    } catch (Throwable e) {
-                        Log.e(TAG, "finish receiver", e);
-                    }
-                }
-            }, intentFilter, null, 0, 0);
-        } catch (Throwable e) {
-            Log.e(TAG, "register boot receiver", e);
-        }
+        var intentFilter = new IntentFilter(Intent.ACTION_LOCKED_BOOT_COMPLETED);
+        intentFilter.setPriority(IntentFilter.SYSTEM_HIGH_PRIORITY);
+        registerReceiver(List.of(intentFilter), 0, this::dispatchBootCompleted);
         Log.d(TAG, "registered boot receiver");
     }
 
     private void registerUserChangeReceiver() {
-        try {
-            IntentFilter userFilter = new IntentFilter();
-            userFilter.addAction(ACTION_USER_ADDED);
-            userFilter.addAction(ACTION_USER_REMOVED);
+        var userFilter = new IntentFilter();
+        userFilter.addAction(ACTION_USER_ADDED);
+        userFilter.addAction(ACTION_USER_REMOVED);
 
-            var receiver = new IIntentReceiver.Stub() {
-                @Override
-                public void performReceive(Intent intent, int resultCode, String data, Bundle extras, boolean ordered, boolean sticky, int sendingUser) {
-                    getExecutorService().submit(() -> dispatchUserChanged(intent));
-                    try {
-                        ActivityManagerService.finishReceiver(this, resultCode, data, extras, false, intent.getFlags());
-                    } catch (Throwable e) {
-                        Log.e(TAG, "finish receiver", e);
-                    }
-                }
-            };
-
-            ActivityManagerService.registerReceiver("android", null, receiver, userFilter, null, -1, 0);
-        } catch (Throwable e) {
-            Log.e(TAG, "register user info change receiver", e);
-        }
+        registerReceiver(List.of(userFilter), -1, this::dispatchUserChanged);
         Log.d(TAG, "registered user info change receiver");
     }
 
+    private void registerOpenManagerReceiver() {
+        var intentFilter = new IntentFilter(LSPNotificationManager.openManagerAction);
+        var moduleFilter = new IntentFilter(intentFilter);
+        moduleFilter.addDataScheme("module");
+
+        registerReceiver(List.of(intentFilter, moduleFilter), 0, this::dispatchOpenManager);
+        Log.d(TAG, "registered open manager receiver");
+    }
+
+    private void registerModuleScopeReceiver() {
+        var intentFilter = new IntentFilter(LSPNotificationManager.moduleScope);
+        intentFilter.addDataScheme("module");
+
+        registerReceiver(List.of(intentFilter), 0, this::dispatchModuleScope);
+        Log.d(TAG, "registered module scope receiver");
+    }
+
+    private void registerUidObserver() {
+        try {
+            var which = HiddenApiBridge.ActivityManager_UID_OBSERVER_ACTIVE()
+                    | HiddenApiBridge.ActivityManager_UID_OBSERVER_GONE()
+                    | HiddenApiBridge.ActivityManager_UID_OBSERVER_IDLE()
+                    | HiddenApiBridge.ActivityManager_UID_OBSERVER_CACHED();
+            LSPModuleService.uidClear();
+            ActivityManagerService.registerUidObserver(new IUidObserver.Stub() {
+                @Override
+                public void onUidActive(int uid) {
+                    LSPModuleService.uidStarts(uid);
+                }
+
+                @Override
+                public void onUidCachedChanged(int uid, boolean cached) {
+                    if (!cached) LSPModuleService.uidStarts(uid);
+                }
+
+                @Override
+                public void onUidIdle(int uid, boolean disabled) {
+                    LSPModuleService.uidStarts(uid);
+                }
+
+                @Override
+                public void onUidGone(int uid, boolean disabled) {
+                    LSPModuleService.uidGone(uid);
+                }
+            }, which, HiddenApiBridge.ActivityManager_PROCESS_STATE_UNKNOWN(), null);
+        } catch (RemoteException e) {
+            Log.e(TAG, "registerUidObserver", e);
+        }
+    }
+
     @Override
-    public void dispatchSystemServerContext(IBinder activityThread, IBinder activityToken, String api) {
+    public void dispatchSystemServerContext(IBinder appThread, IBinder activityToken, String api) {
         Log.d(TAG, "received system context");
+        this.appThread = appThread;
         ConfigManager.getInstance().setApi(api);
-        ActivityManagerService.onSystemServerContext(IApplicationThread.Stub.asInterface(activityThread), activityToken);
+        ActivityManagerService.onSystemServerContext(IApplicationThread.Stub.asInterface(appThread), activityToken);
+        registerBootCompleteReceiver();
         registerPackageReceiver();
-        registerUnlockReceiver();
         registerConfigurationReceiver();
         registerSecretCodeReceiver();
-        registerBootCompleteReceiver();
         registerUserChangeReceiver();
+        registerOpenManagerReceiver();
+        registerModuleScopeReceiver();
+        registerUidObserver();
     }
 
     @Override
     public boolean preStartManager(String pkgName, Intent intent) {
         Log.d(TAG, "checking manager intent");
-        return ServiceManager.getManagerService().preStartManager(pkgName, intent);
+        return ServiceManager.getManagerService().preStartManager(pkgName, intent, false);
     }
 }
