@@ -1,3 +1,22 @@
+/*
+ * This file is part of LSPosed.
+ *
+ * LSPosed is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * LSPosed is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with LSPosed.  If not, see <https://www.gnu.org/licenses/>.
+ *
+ * Copyright (C) 2022 LSPosed Contributors
+ */
+
 package org.lsposed.manager.ui.dialog;
 
 import static org.lsposed.manager.App.TAG;
@@ -36,7 +55,7 @@ public class FlashDialogBuilder extends BlurBehindDialogBuilder {
     private final BorderNestedScrollView rootView;
 
     public FlashDialogBuilder(@NonNull Context context, DialogInterface.OnClickListener cancel) {
-        super(context);
+        super(context, R.style.ThemeOverlay_MaterialAlertDialog_Centered_FullWidthButtons);
         var pref = App.getPreferences();
         var notes = pref.getString("release_notes", "");
         this.zipPath = pref.getString("zip_file", null);
