@@ -23,6 +23,7 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.Menu;
+import android.view.MenuInflater;
 import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
@@ -33,6 +34,8 @@ import androidx.activity.result.contract.ActivityResultContracts;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.appcompat.widget.SearchView;
+import androidx.core.view.MenuProvider;
+import androidx.recyclerview.widget.ConcatAdapter;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
@@ -48,7 +51,7 @@ import org.lsposed.manager.util.ModuleUtil;
 import rikka.material.app.LocaleDelegate;
 import rikka.recyclerview.RecyclerViewKt;
 
-public class AppListFragment extends BaseFragment {
+public class AppListFragment extends BaseFragment implements MenuProvider {
 
     public SearchView searchView;
     private ScopeAdapter scopeAdapter;
@@ -87,7 +90,10 @@ public class AppListFragment extends BaseFragment {
         scopeAdapter = new ScopeAdapter(this, module);
         scopeAdapter.setHasStableIds(true);
         scopeAdapter.registerAdapterDataObserver(observer);
-        binding.recyclerView.setAdapter(scopeAdapter);
+        var concatAdapter = new ConcatAdapter();
+        concatAdapter.addAdapter(scopeAdapter.switchAdaptor);
+        concatAdapter.addAdapter(scopeAdapter);
+        binding.recyclerView.setAdapter(concatAdapter);
         binding.recyclerView.setHasFixedSize(true);
         binding.recyclerView.setLayoutManager(new LinearLayoutManager(requireActivity()));
         binding.recyclerView.getBorderViewDelegate().setBorderVisibilityChangedListener((top, oldTop, bottom, oldBottom) -> binding.appBar.setLifted(!top));
@@ -120,7 +126,9 @@ public class AppListFragment extends BaseFragment {
     public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
         if (module == null) {
-            getNavController().navigate(AppListFragmentDirections.actionAppListFragmentToModulesFragment());
+            if (!safeNavigate(R.id.action_app_list_fragment_to_modules_fragment)) {
+                safeNavigate(R.id.modules_nav);
+            }
         }
     }
 
@@ -132,10 +140,13 @@ public class AppListFragment extends BaseFragment {
         int moduleUserId = args.getModuleUserId();
 
         module = ModuleUtil.getInstance().getModule(modulePackageName, moduleUserId);
-        if (module == null)
-            getNavController().navigate(R.id.action_app_list_fragment_to_modules_fragment);
+        if (module == null) {
+            if (!safeNavigate(R.id.action_app_list_fragment_to_modules_fragment)) {
+                safeNavigate(R.id.modules_nav);
+            }
+        }
 
-        backupLauncher = registerForActivityResult(new ActivityResultContracts.CreateDocument(),
+        backupLauncher = registerForActivityResult(new ActivityResultContracts.CreateDocument("application/gzip"),
                 uri -> {
                     if (uri == null) return;
                     runAsync(() -> {
@@ -182,16 +193,12 @@ public class AppListFragment extends BaseFragment {
     }
 
     @Override
-    public boolean onOptionsItemSelected(@NonNull MenuItem item) {
-        if (scopeAdapter.onOptionsItemSelected(item)) {
-            return true;
-        }
-        return super.onOptionsItemSelected(item);
+    public boolean onMenuItemSelected(@NonNull MenuItem item) {
+        return scopeAdapter.onOptionsItemSelected(item);
     }
 
     @Override
-    public void onPrepareOptionsMenu(@NonNull Menu menu) {
-        super.onPrepareOptionsMenu(menu);
+    public void onPrepareMenu(@NonNull Menu menu) {
         searchView = (SearchView) menu.findItem(R.id.menu_search).getActionView();
         searchView.setOnQueryTextListener(searchListener);
         searchView.addOnAttachStateChangeListener(new View.OnAttachStateChangeListener() {
@@ -208,6 +215,11 @@ public class AppListFragment extends BaseFragment {
         });
         searchView.findViewById(androidx.appcompat.R.id.search_edit_frame).setLayoutDirection(View.LAYOUT_DIRECTION_INHERIT);
         scopeAdapter.onPrepareOptionsMenu(menu);
+    }
+
+    @Override
+    public void onCreateMenu(@NonNull Menu menu, @NonNull MenuInflater menuInflater) {
+
     }
 
     @Override
