@@ -50,6 +50,7 @@ public class ExpandableTextView extends MaterialTextView {
     private final SpannableString collapse;
     private final SpannableString expand;
     private final SpannableStringBuilder sb = new SpannableStringBuilder();
+    private int lineCount = 0;
 
     public ExpandableTextView(Context context) {
         this(context, null);
@@ -90,27 +91,41 @@ public class ExpandableTextView extends MaterialTextView {
 
     @Override
     public boolean onPreDraw() {
-        var lineCount = getLayout().getLineCount();
+        this.getViewTreeObserver().removeOnPreDrawListener(this);
+        if (lineCount == 0) {
+            lineCount = getLayout().getLineCount();
+        }
         if (lineCount > maxLines) {
-            SpannableString s;
-            int end;
+            int hintTextOffsetEnd;
             if (maxLines == getMaxLines()) {
                 nextLines = lineCount + 1;
-                end = getLayout().getLineStart(getMaxLines() - 1);
-                s = expand;
-            } else {
+                hintTextOffsetEnd = getLayout().getLineStart(getMaxLines() - 1);
+                setTextWithSpan(text, hintTextOffsetEnd - 1, expand);
+            } else if (nextLines == getMaxLines()) {
                 nextLines = maxLines;
-                end = text.length();
-                s = collapse;
+                hintTextOffsetEnd = getLayout().getLineStart(getMaxLines() - 1);
+                setTextWithSpan(text, hintTextOffsetEnd, collapse);
             }
-            sb.clearSpans();
-            sb.clear();
-            sb.append(text, 0, end - 1);
-            sb.append("\n");
-            sb.append(s);
-            super.setText(sb, BufferType.NORMAL);
         }
         return super.onPreDraw();
+    }
+
+    private void setTextWithSpan(CharSequence text, int textOffsetEnd,
+                                 SpannableString sbStr) {
+        sb.clearSpans();
+        sb.clear();
+        sb.append(text, 0, textOffsetEnd);
+        sb.append("\n");
+        sb.append(sbStr);
+        super.setText(sb, BufferType.NORMAL);
+    }
+
+    @Override
+    protected void onLayout(boolean changed, int left, int top, int right, int bottom) {
+        super.onLayout(changed, left, top, right, bottom);
+        if (getLayout() != null) {
+            lineCount = getLayout().getLineCount();
+        }
     }
 
     @SuppressLint("ClickableViewAccessibility")
