@@ -29,11 +29,14 @@ import org.lsposed.manager.App;
 import org.lsposed.manager.ConfigManager;
 import org.lsposed.manager.adapters.ScopeAdapter;
 
+import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.util.HashSet;
 import java.util.List;
 import java.util.zip.GZIPInputStream;
 import java.util.zip.GZIPOutputStream;
+
+import rikka.core.os.FileUtils;
 
 public class BackupUtils {
     private static final int VERSION = 2;
@@ -79,10 +82,9 @@ public class BackupUtils {
     public static void restore(Uri uri, String packageName) throws IOException, JSONException {
         try (GZIPInputStream gzipInputStream = new GZIPInputStream(App.getInstance().getContentResolver().openInputStream(uri), 32)) {
             StringBuilder string = new StringBuilder();
-            byte[] data = new byte[32];
-            int bytesRead;
-            while ((bytesRead = gzipInputStream.read(data)) != -1) {
-                string.append(new String(data, 0, bytesRead));
+            try (var os = new ByteArrayOutputStream()) {
+                FileUtils.copy(gzipInputStream, os);
+                string.append(os);
             }
             gzipInputStream.close();
             JSONObject rootObject = new JSONObject(string.toString());
@@ -110,7 +112,7 @@ public class BackupUtils {
                                 scope.add(new ScopeAdapter.ApplicationWithEquals(scopeArray.getString(j), 0));
                             }
                         }
-                        ConfigManager.setModuleScope(name, scope);
+                        ConfigManager.setModuleScope(name, module.legacy, scope);
                     }
                 }
             } else {
