@@ -1,3 +1,22 @@
+/*
+ * This file is part of LSPosed.
+ *
+ * LSPosed is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * LSPosed is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with LSPosed.  If not, see <https://www.gnu.org/licenses/>.
+ *
+ * Copyright (C) 2022 LSPosed Contributors
+ */
+
 package org.lsposed.manager.util;
 
 import android.util.Log;
@@ -16,13 +35,13 @@ import java.io.File;
 import java.io.IOException;
 import java.time.Instant;
 import java.time.ZoneOffset;
+import java.util.Locale;
 
 import okhttp3.Call;
 import okhttp3.Callback;
 import okhttp3.Request;
 import okhttp3.Response;
 import okio.Okio;
-import rikka.material.app.LocaleDelegate;
 
 public class UpdateUtil {
     public static void loadRemoteVersion() {
@@ -42,7 +61,7 @@ public class UpdateUtil {
                     var notes = info.get("body").getAsString();
                     var assetsArray = info.getAsJsonArray("assets");
                     for (var assets : assetsArray) {
-                        checkAssets(assets.getAsJsonObject(), notes, api.toLowerCase(LocaleDelegate.getDefaultLocale()));
+                        checkAssets(assets.getAsJsonObject(), notes, api.toLowerCase(Locale.ROOT));
                     }
                 } catch (Throwable t) {
                     Log.e(App.TAG, t.getMessage(), t);
@@ -69,12 +88,13 @@ public class UpdateUtil {
                 .putInt("latest_version", Integer.parseInt(splitName[2]))
                 .putLong("latest_check", Instant.now().getEpochSecond())
                 .putString("release_notes", releaseNotes)
+                .putString("zip_file", null)
                 .putBoolean("checked", true)
                 .apply();
         var updatedAt = Instant.parse(assets.get("updated_at").getAsString());
         var downloadUrl = assets.get("browser_download_url").getAsString();
-        var nowZipTime = pref.getLong("zip_time", 0);
-        if (updatedAt.isAfter(Instant.ofEpochSecond(nowZipTime))) {
+        var zipTime = pref.getLong("zip_time", 0);
+        if (!updatedAt.equals(Instant.ofEpochSecond(zipTime))) {
             var zip = downloadNewZipSync(downloadUrl, name);
             var size = assets.get("size").getAsLong();
             if (zip != null && zip.length() == size) {
