@@ -20,18 +20,16 @@
 
 package de.robv.android.xposed;
 
-import static org.lsposed.lspd.config.LSPApplicationServiceClient.serviceClient;
+import static org.lsposed.lspd.core.ApplicationServiceClient.serviceClient;
 import static org.lsposed.lspd.deopt.PrebuiltMethodsDeopter.deoptResourceMethods;
 import static de.robv.android.xposed.XposedBridge.hookAllMethods;
-import static de.robv.android.xposed.XposedBridge.sInitPackageResourcesCallbacks;
-import static de.robv.android.xposed.XposedBridge.sInitZygoteCallbacks;
-import static de.robv.android.xposed.XposedBridge.sLoadedPackageCallbacks;
 import static de.robv.android.xposed.XposedHelpers.callMethod;
 import static de.robv.android.xposed.XposedHelpers.findAndHookMethod;
 import static de.robv.android.xposed.XposedHelpers.getObjectField;
 import static de.robv.android.xposed.XposedHelpers.getParameterIndexByType;
 import static de.robv.android.xposed.XposedHelpers.setStaticObjectField;
 
+import android.app.ActivityThread;
 import android.content.pm.ApplicationInfo;
 import android.content.res.Resources;
 import android.content.res.ResourcesImpl;
@@ -40,9 +38,10 @@ import android.content.res.XResources;
 import android.os.Build;
 import android.os.IBinder;
 import android.os.Process;
-import android.util.ArraySet;
+import android.util.ArrayMap;
 import android.util.Log;
 
+import org.lsposed.lspd.impl.LSPosedContext;
 import org.lsposed.lspd.models.PreLoadedApk;
 import org.lsposed.lspd.nativebridge.NativeAPI;
 import org.lsposed.lspd.nativebridge.ResourcesHook;
@@ -53,13 +52,13 @@ import java.lang.ref.WeakReference;
 import java.lang.reflect.Method;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
+import java.util.Optional;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicBoolean;
 
 import de.robv.android.xposed.callbacks.XC_InitPackageResources;
-import de.robv.android.xposed.callbacks.XC_InitZygote;
-import de.robv.android.xposed.callbacks.XC_LoadPackage;
 import de.robv.android.xposed.callbacks.XCallback;
 import hidden.HiddenApiBridge;
 
@@ -86,7 +85,7 @@ public final class XposedInit {
         findAndHookMethod("android.app.ApplicationPackageManager", null, "getResourcesForApplication",
                 ApplicationInfo.class, new XC_MethodHook() {
                     @Override
-                    protected void beforeHookedMethod(MethodHookParam param) {
+                    protected void beforeHookedMethod(MethodHookParam<?> param) {
                         ApplicationInfo app = (ApplicationInfo) param.args[0];
                         XResources.setPackageNameForResDir(app.packageName,
                                 app.uid == Process.myUid() ? app.sourceDir : app.publicSourceDir);
@@ -121,7 +120,7 @@ public final class XposedInit {
         final Class<?> classActivityRes = XposedHelpers.findClassIfExists("android.app.ResourcesManager$ActivityResource", classGTLR.getClassLoader());
         var hooker = new XC_MethodHook() {
             @Override
-            protected void afterHookedMethod(MethodHookParam param) {
+            protected void afterHookedMethod(MethodHookParam<?> param) {
                 // At least on OnePlus 5, the method has an additional parameter compared to AOSP.
                 Object activityToken = null;
                 try {
@@ -147,11 +146,13 @@ public final class XposedInit {
                         //noinspection unchecked
                         resourceReferences = (ArrayList<Object>) getObjectField(param.thisObject, "mResourceReferences");
                     }
-                    if (classActivityRes == null) {
+                    if (activityToken == null || classActivityRes == null) {
                         resourceReferences.add(new WeakReference<>(newRes));
                     } else {
+                        // Android S createResourcesForActivity()
                         var activityRes = XposedHelpers.newInstance(classActivityRes);
                         XposedHelpers.setObjectField(activityRes, "resources", new WeakReference<>(newRes));
+                        resourceReferences.add(activityRes);
                     }
                 }
             }
@@ -164,7 +165,7 @@ public final class XposedInit {
         findAndHookMethod(TypedArray.class, "obtain", Resources.class, int.class,
                 new XC_MethodHook() {
                     @Override
-                    protected void afterHookedMethod(MethodHookParam param) throws Throwable {
+                    protected void afterHookedMethod(MethodHookParam<?> param) throws Throwable {
                         if (param.getResult() instanceof XResources.XTypedArray) {
                             return;
                         }
@@ -184,25 +185,23 @@ public final class XposedInit {
 
         // Replace system resources
         XResources systemRes = new XResources(
-                (ClassLoader) XposedHelpers.getObjectField(Resources.getSystem(), "mClassLoader"));
+                (ClassLoader) XposedHelpers.getObjectField(Resources.getSystem(), "mClassLoader"), null);
         HiddenApiBridge.Resources_setImpl(systemRes, (ResourcesImpl) XposedHelpers.getObjectField(Resources.getSystem(), "mResourcesImpl"));
-        systemRes.initObject(null);
         setStaticObjectField(Resources.class, "mSystem", systemRes);
 
         XResources.init(latestResKey);
     }
 
-    private static XResources cloneToXResources(XC_MethodHook.MethodHookParam param, String resDir) {
+    private static XResources cloneToXResources(XC_MethodHook.MethodHookParam<?> param, String resDir) {
         Object result = param.getResult();
         if (result == null || result instanceof XResources) {
             return null;
         }
 
         // Replace the returned resources with our subclass.
-        XResources newRes = new XResources(
-                (ClassLoader) XposedHelpers.getObjectField(param.getResult(), "mClassLoader"));
+        var newRes = new XResources(
+                (ClassLoader) XposedHelpers.getObjectField(param.getResult(), "mClassLoader"), resDir);
         HiddenApiBridge.Resources_setImpl(newRes, (ResourcesImpl) XposedHelpers.getObjectField(param.getResult(), "mResourcesImpl"));
-        newRes.initObject(resDir);
 
         // Invoke handleInitPackageResources().
         if (newRes.isFirstLoad()) {
@@ -217,79 +216,36 @@ public final class XposedInit {
         return newRes;
     }
 
-    /**
-     * Try to load all modules defined in <code>INSTALLER_DATA_BASE_DIR/conf/modules.list</code>
-     */
-    private static final AtomicBoolean modulesLoaded = new AtomicBoolean(false);
-    private static final Object moduleLoadLock = new Object();
-    // @GuardedBy("moduleLoadLock")
-    private static final ArraySet<String> loadedModules = new ArraySet<>();
+    // only legacy modules have non-empty value
+    private static final Map<String, Optional<String>> loadedModules = new ConcurrentHashMap<>();
 
-    public static ArraySet<String> getLoadedModules() {
-        synchronized (moduleLoadLock) {
-            return loadedModules;
-        }
+    public static Map<String, Optional<String>> getLoadedModules() {
+        return loadedModules;
     }
 
-    public static void loadModules() {
-        boolean hasLoaded = !modulesLoaded.compareAndSet(false, true);
-        if (hasLoaded) {
-            return;
-        }
-        synchronized (moduleLoadLock) {
-            var moduleList = serviceClient.getModulesList();
-            var newLoadedApk = new ArraySet<String>();
-            moduleList.forEach(module -> {
-                var apk = module.apkPath;
-                var name = module.packageName;
-                var file = module.file;
-                if (loadedModules.contains(apk)) {
-                    newLoadedApk.add(apk);
-                } else {
-                    loadedModules.add(apk); // temporarily add it for XSharedPreference
-                    boolean loadSuccess = loadModule(name, apk, file);
-                    if (loadSuccess) {
-                        newLoadedApk.add(apk);
-                    }
-                }
-
-                loadedModules.clear();
-                loadedModules.addAll(newLoadedApk);
-
-                // refresh callback according to current loaded module list
-                pruneCallbacks();
-            });
-        }
+    public static void loadLegacyModules() {
+        var moduleList = serviceClient.getLegacyModulesList();
+        moduleList.forEach(module -> {
+            var apk = module.apkPath;
+            var name = module.packageName;
+            var file = module.file;
+            loadedModules.put(name, Optional.of(apk)); // temporarily add it for XSharedPreference
+            if (!loadModule(name, apk, file)) {
+                loadedModules.remove(name);
+            }
+        });
     }
 
-    // remove deactivated or outdated module callbacks
-    private static void pruneCallbacks() {
-        synchronized (moduleLoadLock) {
-            Object[] loadedPkgSnapshot = sLoadedPackageCallbacks.getSnapshot();
-            Object[] initPkgResSnapshot = sInitPackageResourcesCallbacks.getSnapshot();
-            Object[] initZygoteSnapshot = sInitZygoteCallbacks.getSnapshot();
-            for (Object loadedPkg : loadedPkgSnapshot) {
-                if (loadedPkg instanceof IModuleContext) {
-                    if (!loadedModules.contains(((IModuleContext) loadedPkg).getApkPath())) {
-                        sLoadedPackageCallbacks.remove((XC_LoadPackage) loadedPkg);
-                    }
-                }
+    public static void loadModules(ActivityThread at) {
+        var packages = (ArrayMap<?, ?>) XposedHelpers.getObjectField(at, "mPackages");
+        serviceClient.getModulesList().forEach(module -> {
+            loadedModules.put(module.packageName, Optional.empty());
+            if (!LSPosedContext.loadModule(at, module)) {
+                loadedModules.remove(module.packageName);
+            } else {
+                packages.remove(module.packageName);
             }
-            for (Object initPkgRes : initPkgResSnapshot) {
-                if (initPkgRes instanceof IModuleContext) {
-                    if (!loadedModules.contains(((IModuleContext) initPkgRes).getApkPath())) {
-                        sInitPackageResourcesCallbacks.remove((XC_InitPackageResources) initPkgRes);
-                    }
-                }
-            }
-            for (Object initZygote : initZygoteSnapshot) {
-                if (initZygote instanceof IModuleContext) {
-                    if (!loadedModules.contains(((IModuleContext) initZygote).getApkPath())) {
-                        sInitZygoteCallbacks.remove((XC_InitZygote) initZygote);
-                    }
-                }
-            }
-        }
+        });
     }
 
     /**
@@ -319,23 +275,18 @@ public final class XposedInit {
                     IXposedHookZygoteInit.StartupParam param = new IXposedHookZygoteInit.StartupParam();
                     param.modulePath = apk;
                     param.startsSystemServer = startsSystemServer;
-
-                    XposedBridge.hookInitZygote(new IXposedHookZygoteInit.Wrapper(
-                            (IXposedHookZygoteInit) moduleInstance, param));
                     ((IXposedHookZygoteInit) moduleInstance).initZygote(param);
                     count++;
                 }
 
                 if (moduleInstance instanceof IXposedHookLoadPackage) {
-                    XposedBridge.hookLoadPackage(new IXposedHookLoadPackage.Wrapper(
-                            (IXposedHookLoadPackage) moduleInstance, apk));
+                    XposedBridge.hookLoadPackage(new IXposedHookLoadPackage.Wrapper((IXposedHookLoadPackage) moduleInstance));
                     count++;
                 }
 
                 if (moduleInstance instanceof IXposedHookInitPackageResources) {
                     hookResources();
-                    XposedBridge.hookInitPackageResources(new IXposedHookInitPackageResources.Wrapper(
-                            (IXposedHookInitPackageResources) moduleInstance, apk));
+                    XposedBridge.hookInitPackageResources(new IXposedHookInitPackageResources.Wrapper((IXposedHookInitPackageResources) moduleInstance));
                     count++;
                 }
             } catch (Throwable t) {
@@ -350,7 +301,7 @@ public final class XposedInit {
      * in <code>assets/xposed_init</code>.
      */
     private static boolean loadModule(String name, String apk, PreLoadedApk file) {
-        Log.i(TAG, "Loading module " + name + " from " + apk);
+        Log.i(TAG, "Loading legacy module " + name + " from " + apk);
 
         var sb = new StringBuilder();
         var abis = Process.is64Bit() ? Build.SUPPORTED_64_BIT_ABIS : Build.SUPPORTED_32_BIT_ABIS;
