@@ -21,11 +21,13 @@ package org.lsposed.lspd.service;
 
 import static org.lsposed.lspd.service.ServiceManager.TAG;
 
-import android.annotation.SuppressLint;
+import android.app.ContentProviderHolder;
 import android.app.IActivityManager;
 import android.app.IApplicationThread;
 import android.app.IServiceConnection;
+import android.app.IUidObserver;
 import android.app.ProfilerInfo;
+import android.content.Context;
 import android.content.IContentProvider;
 import android.content.IIntentReceiver;
 import android.content.Intent;
@@ -42,7 +44,7 @@ import android.util.Log;
 public class ActivityManagerService {
     private static IActivityManager am = null;
     private static IBinder binder = null;
-    private static IApplicationThread thread = null;
+    private static IApplicationThread appThread = null;
     private static IBinder token = null;
 
     private static final IBinder.DeathRecipient deathRecipient = new IBinder.DeathRecipient() {
@@ -52,14 +54,14 @@ public class ActivityManagerService {
             binder.unlinkToDeath(this, 0);
             binder = null;
             am = null;
-            thread = null;
+            appThread = null;
             token = null;
         }
     };
 
     public static IActivityManager getActivityManager() {
-        if (binder == null && am == null) {
-            binder = ServiceManager.getService("activity");
+        if (binder == null || am == null) {
+            binder = ServiceManager.getService(Context.ACTIVITY_SERVICE);
             if (binder == null) return null;
             try {
                 binder.linkToDeath(deathRecipient, 0);
@@ -78,15 +80,21 @@ public class ActivityManagerService {
                                                  String resultData, Bundle map, String[] requiredPermissions,
                                                  int appOp, Bundle options, boolean serialized, boolean sticky, int userId) throws RemoteException {
         IActivityManager am = getActivityManager();
-        if (am == null || thread == null) return -1;
+        if (am == null || appThread == null) return -1;
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S) {
-            return am.broadcastIntentWithFeature(thread, callingFeatureId, intent, resolvedType, resultTo,
-                    resultCode, resultData, null, requiredPermissions, null, appOp, null,
-                    serialized, sticky, userId);
-        } else if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) {
-            return am.broadcastIntentWithFeature(thread, callingFeatureId, intent, resolvedType, resultTo, resultCode, resultData, map, requiredPermissions, appOp, options, serialized, sticky, userId);
+            try {
+                return am.broadcastIntentWithFeature(appThread, callingFeatureId, intent, resolvedType, resultTo,
+                        resultCode, resultData, null, requiredPermissions, null, null, appOp, null,
+                        serialized, sticky, userId);
+            } catch (NoSuchMethodError ignored) {
+                return am.broadcastIntentWithFeature(appThread, callingFeatureId, intent, resolvedType, resultTo,
+                        resultCode, resultData, null, requiredPermissions, null, appOp, null,
+                        serialized, sticky, userId);
+            }
+        } else if (Build.VERSION.SDK_INT == Build.VERSION_CODES.R) {
+            return am.broadcastIntentWithFeature(appThread, callingFeatureId, intent, resolvedType, resultTo, resultCode, resultData, map, requiredPermissions, appOp, options, serialized, sticky, userId);
         } else {
-            return am.broadcastIntent(thread, intent, resolvedType, resultTo, resultCode, resultData, map, requiredPermissions, appOp, options, serialized, sticky, userId);
+            return am.broadcastIntent(appThread, intent, resolvedType, resultTo, resultCode, resultData, map, requiredPermissions, appOp, options, serialized, sticky, userId);
         }
     }
 
@@ -102,27 +110,31 @@ public class ActivityManagerService {
         return am.startUserInBackground(userId);
     }
 
-    @SuppressLint("NewApi")
     public static Intent registerReceiver(String callerPackage,
                                           String callingFeatureId, IIntentReceiver receiver, IntentFilter filter,
                                           String requiredPermission, int userId, int flags) throws RemoteException {
         IActivityManager am = getActivityManager();
-        if (am == null || thread == null) return null;
+        if (am == null || appThread == null) return null;
 
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S || (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R && Build.VERSION.PREVIEW_SDK_INT != 0))
-            return am.registerReceiverWithFeature(thread, callerPackage, callingFeatureId, "null", receiver, filter, requiredPermission, userId, flags);
-        else if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) {
-            return am.registerReceiverWithFeature(thread, callerPackage, callingFeatureId, receiver, filter, requiredPermission, userId, flags);
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S)
+            return am.registerReceiverWithFeature(appThread, callerPackage, callingFeatureId, "null", receiver, filter, requiredPermission, userId, flags);
+        else if (Build.VERSION.SDK_INT == Build.VERSION_CODES.R) {
+            return am.registerReceiverWithFeature(appThread, callerPackage, callingFeatureId, receiver, filter, requiredPermission, userId, flags);
         } else {
-            return am.registerReceiver(thread, callerPackage, receiver, filter, requiredPermission, userId, flags);
+            return am.registerReceiver(appThread, callerPackage, receiver, filter, requiredPermission, userId, flags);
         }
     }
 
-    public static void finishReceiver(IBinder who, int resultCode, String resultData, Bundle map,
-                                      boolean abortBroadcast, int flags) throws RemoteException {
+    public static void finishReceiver(IBinder intentReceiver, IBinder applicationThread, int resultCode,
+                                      String resultData, Bundle resultExtras, boolean resultAbort,
+                                      int flags) throws RemoteException {
         IActivityManager am = getActivityManager();
-        if (am == null || thread == null) return;
-        am.finishReceiver(who, resultCode, resultData, map, abortBroadcast, flags);
+        if (am == null || appThread == null) return;
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.UPSIDE_DOWN_CAKE) {
+            am.finishReceiver(applicationThread, resultCode, resultData, resultExtras, resultAbort, flags);
+        } else {
+            am.finishReceiver(intentReceiver, resultCode, resultData, resultExtras, resultAbort, flags);
+        }
     }
 
     public static int bindService(Intent service,
@@ -130,8 +142,11 @@ public class ActivityManagerService {
                                   String callingPackage, int userId) throws RemoteException {
 
         IActivityManager am = getActivityManager();
-        if (am == null || thread == null) return -1;
-        return am.bindService(thread, token, service, resolvedType, connection, flags, callingPackage, userId);
+        if (am == null || appThread == null) return -1;
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.UPSIDE_DOWN_CAKE)
+            return am.bindService(appThread, token, service, resolvedType, connection, (long) flags, callingPackage, userId);
+        else
+            return am.bindService(appThread, token, service, resolvedType, connection, flags, callingPackage, userId);
     }
 
     public static boolean unbindService(IServiceConnection connection) throws RemoteException {
@@ -145,16 +160,17 @@ public class ActivityManagerService {
                                                      IBinder resultTo, String resultWho, int requestCode, int flags,
                                                      ProfilerInfo profilerInfo, Bundle options, int userId) throws RemoteException {
         IActivityManager am = getActivityManager();
-        if (am == null || thread == null) return -1;
+        if (am == null || appThread == null) return -1;
+        intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) {
-            return am.startActivityAsUserWithFeature(thread, callingPackage, callingFeatureId, intent, resolvedType, resultTo, resultWho, requestCode, flags, profilerInfo, options, userId);
+            return am.startActivityAsUserWithFeature(appThread, callingPackage, callingFeatureId, intent, resolvedType, resultTo, resultWho, requestCode, flags, profilerInfo, options, userId);
         } else {
-            return am.startActivityAsUser(thread, callingPackage, intent, resolvedType, resultTo, resultWho, requestCode, flags, profilerInfo, options, userId);
+            return am.startActivityAsUser(appThread, callingPackage, intent, resolvedType, resultTo, resultWho, requestCode, flags, profilerInfo, options, userId);
         }
     }
 
     public static void onSystemServerContext(IApplicationThread thread, IBinder token) {
-        ActivityManagerService.thread = thread;
+        ActivityManagerService.appThread = thread;
         ActivityManagerService.token = token;
     }
 
@@ -179,11 +195,18 @@ public class ActivityManagerService {
     public static IContentProvider getContentProvider(String auth, int userId) throws RemoteException {
         IActivityManager am = getActivityManager();
         if (am == null) return null;
+        ContentProviderHolder holder;
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
-            return am.getContentProviderExternal(auth, userId, token, null).provider;
+            holder = am.getContentProviderExternal(auth, userId, token, null);
         } else {
-            return am.getContentProviderExternal(auth, userId, token).provider;
+            holder = am.getContentProviderExternal(auth, userId, token);
         }
+        return holder != null ? holder.provider : null;
     }
 
+    public static void registerUidObserver(IUidObserver observer, int which, int cutpoint, String callingPackage) throws RemoteException {
+        IActivityManager am = getActivityManager();
+        if (am == null) return;
+        am.registerUidObserver(observer, which, cutpoint, callingPackage);
+    }
 }
