@@ -24,31 +24,18 @@ import static org.lsposed.lspd.service.ServiceManager.TAG;
 import static org.lsposed.lspd.service.ServiceManager.getExecutorService;
 
 import android.annotation.SuppressLint;
-import android.app.INotificationManager;
 import android.app.IServiceConnection;
-import android.app.Notification;
-import android.app.NotificationChannel;
-import android.app.NotificationManager;
-import android.app.PendingIntent;
 import android.content.AttributionSource;
 import android.content.ComponentName;
 import android.content.Context;
 import android.content.Intent;
-import android.content.pm.LauncherApps;
 import android.content.pm.PackageInfo;
 import android.content.pm.PackageManager;
 import android.content.pm.ResolveInfo;
-import android.content.pm.ShortcutInfo;
-import android.content.pm.ShortcutManager;
 import android.content.pm.VersionedPackage;
-import android.graphics.BitmapFactory;
-import android.graphics.Color;
-import android.graphics.drawable.Icon;
 import android.net.Uri;
 import android.os.Build;
 import android.os.Bundle;
-import android.os.Handler;
-import android.os.HandlerThread;
 import android.os.IBinder;
 import android.os.ParcelFileDescriptor;
 import android.os.RemoteException;
@@ -62,47 +49,28 @@ import android.view.IWindowManager;
 import androidx.annotation.NonNull;
 
 import org.lsposed.daemon.BuildConfig;
-import org.lsposed.daemon.R;
 import org.lsposed.lspd.ILSPManagerService;
 import org.lsposed.lspd.models.Application;
 import org.lsposed.lspd.models.UserInfo;
-import org.lsposed.lspd.util.FakeContext;
 import org.lsposed.lspd.util.Utils;
 
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
-import java.lang.reflect.InvocationTargetException;
-import java.util.Collections;
 import java.util.LinkedList;
 import java.util.List;
-import java.util.Map;
 import java.util.UUID;
 import java.util.concurrent.TimeUnit;
-import java.util.zip.ZipFile;
 
 import hidden.HiddenApiBridge;
-import io.github.xposed.xposedservice.utils.ParceledListSlice;
+import io.github.libxposed.service.IXposedService;
+import rikka.parcelablelist.ParcelableListSlice;
 
 public class LSPManagerService extends ILSPManagerService.Stub {
-    private static final String PROP_NAME = "dalvik.vm.dex2oat-flags";
-    private static final String PROP_VALUE = "--inline-max-code-units=0";
     // this maybe useful when obtaining the manager binder
     private static String RANDOM_UUID = null;
-    private static final String SHORTCUT_ID = "org.lsposed.manager.shortcut";
-    public static final int NOTIFICATION_ID = 114514;
-    public static final String CHANNEL_ID = "lsposed";
-    public static final String CHANNEL_NAME = "LSPosed Manager";
-    public static final int CHANNEL_IMP = NotificationManager.IMPORTANCE_HIGH;
 
-    private static final HandlerThread worker = new HandlerThread("manager worker");
-    private static final Handler workerHandler;
     private static Intent managerIntent = null;
-
-    static {
-        worker.start();
-        workerHandler = new Handler(worker.getLooper());
-    }
 
     public class ManagerGuard implements IBinder.DeathRecipient {
         private final @NonNull
@@ -161,97 +129,41 @@ public class LSPManagerService extends ILSPManagerService.Stub {
     LSPManagerService() {
     }
 
-    private static Icon getManagerIcon() {
-        try (var zip = new ZipFile(ConfigFileManager.managerApkPath.toString())) {
-            var entry = zip.getEntry("assets/ic_launcher-playstore.png");
-            try (var is = zip.getInputStream(entry)) {
-                return Icon.createWithAdaptiveBitmap(BitmapFactory.decodeStream(is));
-            }
-        } catch (Throwable e) {
-            Log.e(TAG, "load icon", e);
-        }
-        return null;
-    }
-
-    static Intent getManagerIntent() {
+    private static Intent getManagerIntent() {
+        if (managerIntent != null) return managerIntent;
         try {
-            if (managerIntent == null) {
-                var intent = PackageService.getLaunchIntentForPackage(BuildConfig.MANAGER_INJECTED_PKG_NAME);
-                if (intent == null) {
-                    var pkgInfo = PackageService.getPackageInfo(BuildConfig.MANAGER_INJECTED_PKG_NAME, PackageManager.GET_ACTIVITIES, 0);
-                    if (pkgInfo != null && pkgInfo.activities != null && pkgInfo.activities.length > 0) {
-                        for (var activityInfo : pkgInfo.activities) {
-                            if (activityInfo.processName.equals(activityInfo.packageName)) {
-                                intent = new Intent();
-                                intent.setComponent(new ComponentName(activityInfo.packageName, activityInfo.name));
-                                intent.setAction(Intent.ACTION_MAIN);
-                                break;
-                            }
+            var intent = PackageService.getLaunchIntentForPackage(BuildConfig.MANAGER_INJECTED_PKG_NAME);
+            if (intent == null) {
+                var pkgInfo = PackageService.getPackageInfo(BuildConfig.MANAGER_INJECTED_PKG_NAME, PackageManager.GET_ACTIVITIES, 0);
+                if (pkgInfo != null && pkgInfo.activities != null && pkgInfo.activities.length > 0) {
+                    for (var activityInfo : pkgInfo.activities) {
+                        if (activityInfo.processName.equals(activityInfo.packageName)) {
+                            intent = new Intent();
+                            intent.setComponent(new ComponentName(activityInfo.packageName, activityInfo.name));
+                            intent.setAction(Intent.ACTION_MAIN);
+                            break;
                         }
                     }
                 }
-                if (intent != null) {
-                    if (intent.getCategories() != null) intent.getCategories().clear();
-                    intent.addCategory("org.lsposed.manager.LAUNCH_MANAGER");
-                    intent.setPackage(BuildConfig.MANAGER_INJECTED_PKG_NAME);
-                    managerIntent = (Intent) intent.clone();
-                }
             }
-        } catch (Throwable e) {
+            if (intent != null) {
+                if (intent.getCategories() != null) intent.getCategories().clear();
+                intent.addCategory("org.lsposed.manager.LAUNCH_MANAGER");
+                intent.setPackage(BuildConfig.MANAGER_INJECTED_PKG_NAME);
+                managerIntent = new Intent(intent);
+            }
+        } catch (RemoteException e) {
             Log.e(TAG, "get Intent", e);
         }
         return managerIntent;
-
     }
 
-    public static PendingIntent getNotificationIntent(String modulePackageName, int moduleUserId) {
-        try {
-            var intent = (Intent) getManagerIntent().clone();
-            intent.setData(new Uri.Builder().scheme("module").encodedAuthority(modulePackageName + ":" + moduleUserId).build());
-            return PendingIntent.getActivity(new FakeContext(), 0, intent, PendingIntent.FLAG_UPDATE_CURRENT | PendingIntent.FLAG_IMMUTABLE);
-        } catch (Throwable e) {
-            Log.e(TAG, "get notification intent", e);
-            return null;
-        }
-    }
-
-    public static void showNotification(String modulePackageName,
-                                        int moduleUserId,
-                                        boolean enabled,
-                                        boolean systemModule) {
-        try {
-            var context = new FakeContext();
-            String title = context.getString(enabled ? systemModule ?
-                    R.string.xposed_module_updated_notification_title_system :
-                    R.string.xposed_module_updated_notification_title :
-                    R.string.module_is_not_activated_yet);
-            String content = context.getString(enabled ? systemModule ?
-                    R.string.xposed_module_updated_notification_content_system :
-                    R.string.xposed_module_updated_notification_content :
-                    R.string.module_is_not_activated_yet_detailed, modulePackageName);
-
-            var style = new Notification.BigTextStyle();
-            style.bigText(content);
-
-            var notification = new Notification.Builder(context, CHANNEL_ID)
-                    .setContentTitle(title)
-                    .setContentText(content)
-                    .setSmallIcon(android.R.drawable.ic_dialog_info)
-                    .setColor(Color.BLUE)
-                    .setContentIntent(getNotificationIntent(modulePackageName, moduleUserId))
-                    .setAutoCancel(true)
-                    .setStyle(style)
-                    .build();
-            notification.extras.putString("android.substName", "LSPosed");
-            var im = INotificationManager.Stub.asInterface(android.os.ServiceManager.getService("notification"));
-            final NotificationChannel channel =
-                    new NotificationChannel(CHANNEL_ID, CHANNEL_NAME, CHANNEL_IMP);
-            im.createNotificationChannels("android",
-                    new android.content.pm.ParceledListSlice<>(Collections.singletonList(channel)));
-            im.enqueueNotificationWithTag("android", "android", "114514", NOTIFICATION_ID, notification, 0);
-        } catch (Throwable e) {
-            Log.e(TAG, "post notification", e);
-        }
+    static void openManager(Uri withData) {
+        var intent = getManagerIntent();
+        if (intent == null) return;
+        intent = new Intent(intent);
+        intent.setData(withData);
+        ServiceManager.getManagerService().preStartManager(BuildConfig.MANAGER_INJECTED_PKG_NAME, intent, true);
     }
 
     @SuppressLint("WrongConstant")
@@ -271,72 +183,8 @@ public class LSPManagerService extends ILSPManagerService.Stub {
                     null, null, 0, null, null,
                     null, -1, null, true, false,
                     0);
-        } catch (Throwable t) {
+        } catch (RemoteException t) {
             Log.e(TAG, "Broadcast to manager failed: ", t);
-        }
-    }
-
-    public static void createOrUpdateShortcut(boolean force) {
-        workerHandler.post(() -> createOrUpdateShortcutInternal(force, true));
-    }
-
-    public static void createOrUpdateShortcut(boolean force, boolean shouldCreate) {
-        workerHandler.post(() -> createOrUpdateShortcutInternal(force, shouldCreate));
-    }
-
-    private synchronized static void createOrUpdateShortcutInternal(boolean force, boolean shouldCreate) {
-        try {
-            while (!UserService.isUserUnlocked(0)) {
-                Log.d(TAG, "user is not yet unlocked, waiting for 1s...");
-                Thread.sleep(1000);
-            }
-            var smCtor = ShortcutManager.class.getDeclaredConstructor(Context.class);
-            smCtor.setAccessible(true);
-            var context = new FakeContext("com.android.settings");
-            var sm = smCtor.newInstance(context);
-            if (!sm.isRequestPinShortcutSupported()) {
-                Log.d(TAG, "pinned shortcut not supported, skipping");
-                return;
-            }
-            var intent = getManagerIntent();
-            var settingIntent = PackageService.getLaunchIntentForPackage("com.android.settings");
-            var componentName = settingIntent != null ? settingIntent.getComponent() : new ComponentName("com.android.settings", "android.__dummy__");
-            var shortcut = new ShortcutInfo.Builder(context, SHORTCUT_ID)
-                    .setShortLabel("LSPosed")
-                    .setLongLabel("LSPosed")
-                    .setIntent(intent)
-                    .setActivity(componentName)
-                    .setCategories(intent.getCategories())
-                    .setIcon(getManagerIcon())
-                    .build();
-
-            for (var shortcutInfo : sm.getPinnedShortcuts()) {
-                if (SHORTCUT_ID.equals(shortcutInfo.getId()) && shortcutInfo.isPinned()) {
-                    var shortcutIntent = sm.createShortcutResultIntent(shortcutInfo);
-                    var request = (LauncherApps.PinItemRequest) shortcutIntent.getParcelableExtra(LauncherApps.EXTRA_PIN_ITEM_REQUEST);
-                    var requestInfo = request.getShortcutInfo();
-                    // https://cs.android.com/android/platform/superproject/+/android-8.1.0_r1:frameworks/base/services/core/java/com/android/server/pm/ShortcutRequestPinProcessor.java;drc=4ad6b57700bef4c484021f49e018117046562e6b;l=337
-                    if (requestInfo.isPinned()) {
-                        Log.d(TAG, "shortcut exists, updating");
-                        sm.updateShortcuts(Collections.singletonList(shortcut));
-                        return;
-                    }
-                }
-            }
-            var configManager = ConfigManager.getInstance();
-            if (!force && configManager.isManagerInstalled()) {
-                Log.d(TAG, "Manager has installed, skip adding shortcut");
-                return;
-            }
-            // Only existing shortcuts are updated when system settings
-            // are changed and no new shortcuts are requested
-            if (!force && !shouldCreate) return;
-            if (configManager.isAddShortcut()) {
-                sm.requestPinShortcut(shortcut, null);
-                Log.d(TAG, "done add shortcut");
-            }
-        } catch (Throwable e) {
-            Log.e(TAG, "add shortcut", e);
         }
     }
 
@@ -347,7 +195,7 @@ public class LSPManagerService extends ILSPManagerService.Stub {
 
     private void ensureWebViewPermission(File f) {
         if (!f.exists()) return;
-        SELinux.setFileContext(f.getAbsolutePath(), "u:object_r:privapp_data_file:s0");
+        SELinux.setFileContext(f.getAbsolutePath(), "u:object_r:magisk_file:s0");
         try {
             Os.chown(f.getAbsolutePath(), BuildConfig.MANAGER_INJECTED_UID, BuildConfig.MANAGER_INJECTED_UID);
         } catch (ErrnoException e) {
@@ -363,16 +211,12 @@ public class LSPManagerService extends ILSPManagerService.Stub {
     private void ensureWebViewPermission() {
         try {
             var pkgInfo = PackageService.getPackageInfo(BuildConfig.MANAGER_INJECTED_PKG_NAME, 0, 0);
-            File cacheDir = null;
             if (pkgInfo != null) {
-                cacheDir = new File(HiddenApiBridge.ApplicationInfo_credentialProtectedDataDir(pkgInfo.applicationInfo) + "/cache");
+                var cacheDir = new File(HiddenApiBridge.ApplicationInfo_credentialProtectedDataDir(pkgInfo.applicationInfo) + "/cache");
+                // The cache directory does not exist after `pm clear`
+                cacheDir.mkdirs();
+                ensureWebViewPermission(cacheDir);
             }
-            var webviewDir = new File(cacheDir, "WebView");
-            webviewDir.mkdirs();
-            var httpCacheDir = new File(cacheDir, "http_cache");
-            httpCacheDir.mkdirs();
-            ensureWebViewPermission(webviewDir);
-            ensureWebViewPermission(httpCacheDir);
         } catch (Throwable e) {
             Log.w(TAG, "cannot ensure webview dir", e);
         }
@@ -392,7 +236,7 @@ public class LSPManagerService extends ILSPManagerService.Stub {
     // where starting the target app while the manager is
     // still running.
     // We instead let the manager to restart the activity.
-    synchronized boolean preStartManager(String pkgName, Intent intent) {
+    synchronized boolean preStartManager(String pkgName, Intent intent, boolean doResume) {
         // first, check if it's our target app, if not continue the start
         if (BuildConfig.MANAGER_INJECTED_PKG_NAME.equals(pkgName)) {
             Log.d(TAG, "starting target app of parasitic manager");
@@ -409,7 +253,20 @@ public class LSPManagerService extends ILSPManagerService.Stub {
                     Log.d(TAG, "manager is still running or is on its way");
                     // there's one running parasitic manager
                     // or it's run by ourself after killing, resume it
+                    if (doResume) {
+                        // if doResume is true, we help do the resumption
+                        try {
+                            ActivityManagerService.startActivityAsUserWithFeature("android", null, intent, intent.getType(), null, null, 0, 0, null, null, 0);
+                        } catch (Throwable e) {
+                            Log.w(TAG, "resume manager", e);
+                        }
+                    }
                     return true;
+                } else if (pendingManager) {
+                    // Check the flag in case new launch comes before finishing
+                    // the previous one to avoid racing.
+                    Log.d(TAG, "manager is still on its way when new launch comes, skipping");
+                    return false;
                 } else {
                     // new parasitic manager launch, set the flag and kill
                     // old processes
@@ -444,7 +301,7 @@ public class LSPManagerService extends ILSPManagerService.Stub {
             ActivityManagerService.forceStopPackage(packageName, 0);
             Log.d(TAG, "stopped old package");
             if (addUUID) {
-                intent = (Intent) intent.clone();
+                intent = new Intent(intent);
                 RANDOM_UUID = UUID.randomUUID().toString();
                 intent.addCategory(RANDOM_UUID);
             }
@@ -505,7 +362,7 @@ public class LSPManagerService extends ILSPManagerService.Stub {
 
     @Override
     public int getXposedApiVersion() {
-        return BuildConfig.API_CODE;
+        return IXposedService.API;
     }
 
     @Override
@@ -524,7 +381,7 @@ public class LSPManagerService extends ILSPManagerService.Stub {
     }
 
     @Override
-    public ParceledListSlice<PackageInfo> getInstalledPackagesFromAllUsers(int flags, boolean filterNoProcess) throws RemoteException {
+    public ParcelableListSlice<PackageInfo> getInstalledPackagesFromAllUsers(int flags, boolean filterNoProcess) throws RemoteException {
         return PackageService.getInstalledPackagesFromAllUsers(flags, filterNoProcess);
     }
 
@@ -539,31 +396,18 @@ public class LSPManagerService extends ILSPManagerService.Stub {
     }
 
     @Override
-    public boolean setModuleScope(String packageName, ParceledListSlice<Application> scope) throws RemoteException {
-        return ConfigManager.getInstance().setModuleScope(packageName, scope.getList());
+    public boolean setModuleScope(String packageName, List<Application> scope) throws RemoteException {
+        return ConfigManager.getInstance().setModuleScope(packageName, scope);
     }
 
     @Override
-    public ParceledListSlice<Application> getModuleScope(String packageName) {
-        List<Application> list = ConfigManager.getInstance().getModuleScope(packageName);
-        if (list == null) return null;
-        else return new ParceledListSlice<>(list);
+    public List<Application> getModuleScope(String packageName) {
+        return ConfigManager.getInstance().getModuleScope(packageName);
     }
 
     @Override
     public boolean disableModule(String packageName) {
         return ConfigManager.getInstance().disableModule(packageName);
-    }
-
-    @Override
-    public boolean isAddShortcut() {
-        return ConfigManager.getInstance().isAddShortcut();
-    }
-
-    @Override
-    public void setAddShortcut(boolean enabled) {
-        ConfigManager.getInstance().setAddShortcut(enabled);
-        if (enabled) createOrUpdateShortcut(true);
     }
 
     @Override
@@ -583,7 +427,7 @@ public class LSPManagerService extends ILSPManagerService.Stub {
 
     @Override
     public ParcelFileDescriptor getModulesLog() {
-        workerHandler.post(() -> ServiceManager.getLogcatService().checkLogFile());
+        ServiceManager.getLogcatService().checkLogFile();
         return ConfigManager.getInstance().getModulesLog();
     }
 
@@ -603,18 +447,20 @@ public class LSPManagerService extends ILSPManagerService.Stub {
     }
 
     @Override
-    public void reboot(boolean shutdown) {
-        var value = shutdown ? "shutdown" : "reboot";
-        SystemProperties.set("sys.powerctl", value);
+    public void reboot() throws RemoteException {
+        PowerService.reboot(false, null, false);
     }
 
     @Override
     public boolean uninstallPackage(String packageName, int userId) throws RemoteException {
         try {
-            if (ActivityManagerService.startUserInBackground(userId))
-                return PackageService.uninstallPackage(new VersionedPackage(packageName, PackageManager.VERSION_CODE_HIGHEST), userId);
-            else return false;
-        } catch (InterruptedException | InvocationTargetException | NoSuchMethodException | InstantiationException | IllegalAccessException e) {
+            if (ActivityManagerService.startUserInBackground(userId)) {
+                var pkg = new VersionedPackage(packageName, PackageManager.VERSION_CODE_HIGHEST);
+                return PackageService.uninstallPackage(pkg, userId);
+            } else {
+                return false;
+            }
+        } catch (InterruptedException | ReflectiveOperationException e) {
             Log.e(TAG, e.getMessage(), e);
             return false;
         }
@@ -622,7 +468,8 @@ public class LSPManagerService extends ILSPManagerService.Stub {
 
     @Override
     public boolean isSepolicyLoaded() {
-        return ConfigManager.getInstance().isSepolicyLoaded();
+        return SELinux.checkSELinuxAccess("u:r:dex2oat:s0", "u:object_r:dex2oat_exec:s0",
+                "file", "execute_no_trans");
     }
 
     @Override
@@ -664,7 +511,7 @@ public class LSPManagerService extends ILSPManagerService.Stub {
             if (parent < 0) return -1;
             if (currentUser.id != parent) {
                 if (!ActivityManagerService.switchUser(parent)) return -1;
-                var window = android.os.ServiceManager.getService("window");
+                var window = android.os.ServiceManager.getService(Context.WINDOW_SERVICE);
                 if (window != null) {
                     var wm = IWindowManager.Stub.asInterface(window);
                     wm.lockNow(null);
@@ -675,13 +522,13 @@ public class LSPManagerService extends ILSPManagerService.Stub {
     }
 
     @Override
-    public ParceledListSlice<ResolveInfo> queryIntentActivitiesAsUser(Intent intent, int flags, int userId) throws RemoteException {
+    public ParcelableListSlice<ResolveInfo> queryIntentActivitiesAsUser(Intent intent, int flags, int userId) throws RemoteException {
         return PackageService.queryIntentActivities(intent, intent.getType(), flags, userId);
     }
 
     @Override
     public boolean dex2oatFlagsLoaded() {
-        return SystemProperties.get(PROP_NAME).contains(PROP_VALUE);
+        return SystemProperties.get("dalvik.vm.dex2oat-flags").contains("--inline-max-code-units=0");
     }
 
     @Override
@@ -696,9 +543,9 @@ public class LSPManagerService extends ILSPManagerService.Stub {
                     if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S) {
                         contentProvider.call(new AttributionSource.Builder(1000).setPackageName("android").build(),
                                 "settings", "PUT_global", "show_hidden_icon_apps_enabled", args);
-                    } else if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) {
+                    } else if (Build.VERSION.SDK_INT == Build.VERSION_CODES.R) {
                         contentProvider.call("android", null, "settings", "PUT_global", "show_hidden_icon_apps_enabled", args);
-                    } else if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
+                    } else if (Build.VERSION.SDK_INT == Build.VERSION_CODES.Q) {
                         contentProvider.call("android", "settings", "PUT_global", "show_hidden_icon_apps_enabled", args);
                     }
                 } catch (NoSuchMethodError e) {
@@ -711,20 +558,14 @@ public class LSPManagerService extends ILSPManagerService.Stub {
     }
 
     @Override
-    public Map<String, ParcelFileDescriptor> getLogs() {
-        return ConfigFileManager.getLogs();
+    public void getLogs(ParcelFileDescriptor zipFd) {
+        ConfigFileManager.getLogs(zipFd);
     }
 
     @Override
     public void restartFor(Intent intent) throws RemoteException {
         forceStopPackage(BuildConfig.MANAGER_INJECTED_PKG_NAME, 0);
         stopAndStartActivity(BuildConfig.MANAGER_INJECTED_PKG_NAME, intent, false);
-    }
-
-    @Override
-    public void createShortcut() {
-        createOrUpdateShortcut(true);
-        setAddShortcut(true);
     }
 
     @Override
@@ -744,7 +585,7 @@ public class LSPManagerService extends ILSPManagerService.Stub {
                 if (exit == 0) {
                     fdw.write("- Reboot after 5s\n".getBytes());
                     Thread.sleep(5000);
-                    reboot(false);
+                    reboot();
                 } else {
                     var s = "! Flash failed, exit with " + exit + "\n";
                     fdw.write(s.getBytes());
@@ -753,13 +594,52 @@ public class LSPManagerService extends ILSPManagerService.Stub {
                 proc.destroy();
                 fdw.write("! Timeout, abort\n".getBytes());
             }
-        } catch (IOException | InterruptedException e) {
+        } catch (IOException | InterruptedException | RemoteException e) {
             Log.e(TAG, "flashZip: ", e);
+        }
+    }
+
+    @Override
+    public void clearApplicationProfileData(String packageName) throws RemoteException {
+        PackageService.clearApplicationProfileData(packageName);
+    }
+
+    @Override
+    public boolean enableStatusNotification() {
+        return ConfigManager.getInstance().enableStatusNotification();
+    }
+
+    @Override
+    public void setEnableStatusNotification(boolean enable) {
+        ConfigManager.getInstance().setEnableStatusNotification(enable);
+        if (enable) {
+            LSPNotificationManager.notifyStatusNotification();
+        } else {
+            LSPNotificationManager.cancelStatusNotification();
         }
     }
 
     @Override
     public boolean performDexOptMode(String packageName) throws RemoteException {
         return PackageService.performDexOptMode(packageName);
+    }
+
+    @Override
+    public boolean getDexObfuscate() {
+        return ConfigManager.getInstance().dexObfuscate();
+    }
+
+    @Override
+    public void setDexObfuscate(boolean enabled) {
+        ConfigManager.getInstance().setDexObfuscate(enabled);
+    }
+
+    @Override
+    public int getDex2OatWrapperCompatibility() {
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
+            return ServiceManager.getDex2OatService().getCompatibility();
+        } else {
+            return 0;
+        }
     }
 }
