@@ -1,5 +1,5 @@
 /*
- * <!--This file is part of LSPosed.
+ * This file is part of LSPosed.
  *
  * LSPosed is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU General Public License
  * along with LSPosed.  If not, see <https://www.gnu.org/licenses/>.
  *
- * Copyright (C) 2021 LSPosed Contributors-->
+ * Copyright (C) 2021 - 2022 LSPosed Contributors
  */
 
 package org.lsposed.lspd.service;
@@ -47,11 +47,15 @@ public class LSPSystemServerService extends ILSPSystemServerService.Stub impleme
     }
 
     public LSPSystemServerService(int maxRetry) {
+        Log.d(TAG, "LSPSystemServerService::LSPSystemServerService");
         requested = -maxRetry;
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) {
+            // Registers a callback when system is registering an authentic "serial" service
+            // And we are proxying all requests to that system service
             var serviceCallback = new IServiceCallback.Stub() {
                 @Override
                 public void onRegistration(String name, IBinder binder) {
+                    Log.d(TAG, "LSPSystemServerService::LSPSystemServerService onRegistration: " + name + " " + binder);
                     if (name.equals(PROXY_SERVICE_NAME) && binder != null && binder != LSPSystemServerService.this) {
                         Log.d(TAG, "Register " + name + " " + binder);
                         originService = binder;
@@ -74,20 +78,46 @@ public class LSPSystemServerService extends ILSPSystemServerService.Stub impleme
 
     @Override
     public ILSPApplicationService requestApplicationService(int uid, int pid, String processName, IBinder heartBeat) {
+        Log.d(TAG, "ILSPApplicationService.requestApplicationService: " + uid + " " + pid + " " + processName + " " + heartBeat);
         requested = 1;
-        if (ConfigManager.getInstance().shouldSkipSystemServer() || uid != 1000 || heartBeat == null || !"android".equals(processName))
+        if (ConfigManager.getInstance().shouldSkipSystemServer() || uid != 1000 || heartBeat == null || !"system".equals(processName))
             return null;
         else
-            return ServiceManager.requestApplicationService(uid, pid, heartBeat);
+            return ServiceManager.requestApplicationService(uid, pid, processName, heartBeat);
     }
 
     @Override
     public boolean onTransact(int code, Parcel data, Parcel reply, int flags) throws RemoteException {
+        Log.d(TAG, "LSPSystemServerService.onTransact: code=" + code);
         if (originService != null) {
             return originService.transact(code, data, reply, flags);
         }
 
-        return super.onTransact(code, data, reply, flags);
+        switch (code) {
+            case BridgeService.TRANSACTION_CODE -> {
+                int uid = data.readInt();
+                int pid = data.readInt();
+                String processName = data.readString();
+                IBinder heartBeat = data.readStrongBinder();
+                var service = requestApplicationService(uid, pid, processName, heartBeat);
+                if (service != null) {
+                    Log.d(TAG, "LSPSystemServerService.onTransact requestApplicationService granted: " + service);
+                    reply.writeNoException();
+                    reply.writeStrongBinder(service.asBinder());
+                    return true;
+                } else {
+                    Log.d(TAG, "LSPSystemServerService.onTransact requestApplicationService rejected");
+                    return false;
+                }
+            }
+            case LSPApplicationService.OBFUSCATION_MAP_TRANSACTION_CODE, LSPApplicationService.DEX_TRANSACTION_CODE -> {
+                // Proxy LSP dex transaction to Application Binder
+                return ServiceManager.getApplicationService().onTransact(code, data, reply, flags);
+            }
+            default -> {
+                return super.onTransact(code, data, reply, flags);
+            }
+        }
     }
 
     public void linkToDeath() {
