@@ -19,20 +19,33 @@
 
 package org.lsposed.lspd.service;
 
+import android.app.ActivityThread;
 import android.content.Context;
 import android.ddm.DdmHandleAppName;
+import android.os.Binder;
+import android.os.Build;
 import android.os.IBinder;
 import android.os.IServiceManager;
 import android.os.Looper;
+import android.os.Parcel;
 import android.os.Process;
+import android.os.RemoteException;
 import android.util.Log;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.annotation.RequiresApi;
 
 import com.android.internal.os.BinderInternal;
 
 import org.lsposed.daemon.BuildConfig;
 
 import java.io.File;
-import java.util.concurrent.ConcurrentHashMap;
+import java.lang.reflect.Field;
+import java.lang.reflect.InvocationTargetException;
+import java.lang.reflect.Method;
+import java.util.List;
+import java.util.Map;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 
@@ -40,7 +53,6 @@ import hidden.HiddenApiBridge;
 
 public class ServiceManager {
     public static final String TAG = "LSPosedService";
-    private static final ConcurrentHashMap<String, LSPModuleService> moduleServices = new ConcurrentHashMap<>();
     private static final File globalNamespace = new File("/proc/1/root");
     @SuppressWarnings("FieldCanBeLocal")
     private static LSPosedService mainService = null;
@@ -48,8 +60,14 @@ public class ServiceManager {
     private static LSPManagerService managerService = null;
     private static LSPSystemServerService systemServerService = null;
     private static LogcatService logcatService = null;
+    private static Dex2OatService dex2OatService = null;
 
-    private static final ExecutorService executorService = Executors.newCachedThreadPool();
+    private static final ExecutorService executorService = Executors.newSingleThreadExecutor();
+
+    @RequiresApi(Build.VERSION_CODES.Q)
+    public static Dex2OatService getDex2OatService() {
+        return dex2OatService;
+    }
 
     public static ExecutorService getExecutorService() {
         return executorService;
@@ -97,21 +115,32 @@ public class ServiceManager {
         logcatService = new LogcatService();
         logcatService.start();
 
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R)
+            permissionManagerWorkaround();
+
         Process.setThreadPriority(Process.THREAD_PRIORITY_FOREGROUND);
         Looper.prepareMainLooper();
+
+
         mainService = new LSPosedService();
         applicationService = new LSPApplicationService();
         managerService = new LSPManagerService();
         systemServerService = new LSPSystemServerService(systemServerMaxRetry);
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
+            dex2OatService = new Dex2OatService();
+            dex2OatService.start();
+        }
 
         systemServerService.putBinderForSystemServer();
-
-        DdmHandleAppName.setAppName("lspd", 0);
 
         // get config before package service is started
         // otherwise getInstance will trigger module/scope cache
         var configManager = ConfigManager.getInstance();
         // --- DO NOT call ConfigManager.getInstance later!!! ---
+
+        ActivityThread.systemMain();
+
+        DdmHandleAppName.setAppName("org.lsposed.daemon", 0);
 
         waitSystemService("package");
         waitSystemService("activity");
@@ -153,16 +182,12 @@ public class ServiceManager {
         throw new RuntimeException("Main thread loop unexpectedly exited");
     }
 
-    public static LSPModuleService getModuleService(String module) {
-        return moduleServices.computeIfAbsent(module, LSPModuleService::new);
-    }
-
     public static LSPApplicationService getApplicationService() {
         return applicationService;
     }
 
-    public static LSPApplicationService requestApplicationService(int uid, int pid, IBinder heartBeat) {
-        if (applicationService.registerHeartBeat(uid, pid, heartBeat))
+    public static LSPApplicationService requestApplicationService(int uid, int pid, String processName, IBinder heartBeat) {
+        if (applicationService.registerHeartBeat(uid, pid, processName, heartBeat))
             return applicationService;
         else return null;
     }
@@ -195,5 +220,57 @@ public class ServiceManager {
 
     public static boolean existsInGlobalNamespace(String path) {
         return toGlobalNamespace(path).exists();
+    }
+
+    private static void permissionManagerWorkaround() {
+        try {
+            Field sCacheField = android.os.ServiceManager.class.getDeclaredField("sCache");
+            sCacheField.setAccessible(true);
+            var sCache = (Map) sCacheField.get(null);
+            sCache.put("permissionmgr", new BinderProxy("permissionmgr"));
+            sCache.put("legacy_permission", new BinderProxy("legacy_permission"));
+            sCache.put("appops", new BinderProxy("appops"));
+        } catch (Throwable e) {
+            Log.e(TAG, "failed to init permission manager", e);
+        }
+    }
+
+    private static class BinderProxy extends Binder {
+        private static final Method rawGetService;
+
+        static {
+            try {
+                rawGetService = android.os.ServiceManager.class.getDeclaredMethod("rawGetService", String.class);
+                rawGetService.setAccessible(true);
+            } catch (NoSuchMethodException e) {
+                throw new RuntimeException(e);
+            }
+        }
+        private IBinder mReal = null;
+        private final String mName;
+
+        BinderProxy(String name) {
+            mName = name;
+        }
+
+        @Override
+        protected boolean onTransact(int code, @NonNull Parcel data, @Nullable Parcel reply, int flags) throws RemoteException {
+            synchronized (this) {
+                if (mReal == null) {
+                    try {
+                        mReal = (IBinder) rawGetService.invoke(null, mName);
+                    } catch (IllegalAccessException | InvocationTargetException ignored){
+
+                    }
+                }
+                if (mReal != null) {
+                    return mReal.transact(code, data, reply, flags);
+                }
+            }
+            // getSplitPermissions
+            if (reply != null && mName.equals("permissionmgr"))
+                reply.writeTypedList(List.of());
+            return true;
+        }
     }
 }
